Polish franchize action boundary docs
