+++ app/franchize/actions.ts
@@ -1,2600 +1,37 @@
 "use server";
 
-import { createInvoice, supabaseAdmin } from "@/lib/supabase-server";
-import { notifyAdmin, sendTelegramDocument, sendTelegramInvoice } from "@/app/actions";
-import { logger } from "@/lib/logger";
-import { z } from "zod";
-import { randomUUID, webcrypto } from "crypto";
-import { readFile } from "fs/promises";
-import path from "path";
-import { getCrewSensitiveData, getUserSensitiveData, saveCrewSensitiveData } from "@/app/lib/private-secrets";
-import { buildFranchizeDocxFromTemplate } from "@/app/franchize/lib/docx-capability";
-import { resolveFranchizeTheme, resolvePaletteByMode } from "@/app/franchize/lib/theme-resolver";
-import type { FranchizeTheme } from "@/lib/franchize-config";
-import {
-  DEFAULT_AD_CARDS_TEXT,
-  DEFAULT_CATEGORY_ORDER,
-  DEFAULT_CONTRACT_PREFILL,
-  DEFAULT_DELIVERY_MODES_TEXT,
-  DEFAULT_FOOTER_TEXT_COLOR,
-  DEFAULT_FRANCHIZE_BRAND,
-  DEFAULT_FRANCHIZE_THEME,
-  DEFAULT_LIGHT_THEME_PALETTE,
-  DEFAULT_MAP_BOUNDS,
-  DEFAULT_MAP_GPS,
-  DEFAULT_MAP_IMAGE_URL,
-  DEFAULT_MENU_LINK_TEMPLATES,
-  DEFAULT_PAYMENT_OPTIONS_TEXT,
-  DEFAULT_PROMO_BANNERS_TEXT,
-  DEFAULT_SHOWCASE_GROUPS,
-  DEFAULT_SOCIAL_LINKS_TEXT,
-  DEFAULT_TELEGRAM_BOT_URL,
-} from "@/lib/franchize-config";
-
-type UnknownRecord = Record<string, unknown>;
-type RentalAvailabilityRow = {
-  vehicle_id: string | null;
-  status: string | null;
-  agreed_start_date: string | null;
-  agreed_end_date: string | null;
-  requested_end_date: string | null;
-};
-
-export type { FranchizeTheme } from "@/lib/franchize-config";
-
-export interface FranchizeHeaderVM {
-  brandName: string;
-  tagline: string;
-  logoUrl: string;
-  logoHref: string;
-  menuLinks: Array<{ label: string; href: string }>;
-}
-
-export interface RentalReviewVM {
-  id: string;
-  rentalId: string;
-  userId: string;
-  bikeId: string;
-  crewId: string;
-  rating: number;
-  text: string;
-  createdAt: string;
-  hiddenAt?: string | null;
-}
-
-export interface RentalReviewSummaryVM {
-  average: number;
-  count: number;
-  latest?: RentalReviewVM;
-  reviews: RentalReviewVM[];
-}
-
-export interface CatalogItemVM {
-  id: string;
-  title: string;
-  subtitle: string;
-  description: string;
-  imageUrl: string;
-  mediaUrls: string[];
-  pricePerDay: number;
-  rentPriceLabel: string;
-  category: string;
-  availabilityStatus: "available" | "busy";
-  availabilityLabel: string;
-  isHot: boolean;
-  saleAvailable: boolean;
-  salePrice: number | null;
-  specs: Array<{ label: string; value: string }>;
-  rawSpecs?: Record<string, unknown>;
-  reviewSummary: RentalReviewSummaryVM;
-}
-
-export interface FranchizeCrewVM {
-  id: string;
-  slug: string;
-  name: string;
-  description: string;
-  logoUrl: string;
-  hqLocation: string;
-  isFound: boolean;
-  theme: FranchizeTheme;
-  header: FranchizeHeaderVM;
-  contacts: {
-    phone: string;
-    email: string;
-    address: string;
-    telegram: string;
-    workingHours: string;
-    map: {
-      gps: string;
-      publicTransport: string;
-      carDirections: string;
-      imageUrl: string;
-      bounds: {
-        top: number;
-        bottom: number;
-        left: number;
-        right: number;
-      };
-    };
-  };
-  catalog: {
-    categories: string[];
-    quickLinks: string[];
-    tickerItems: Array<{ id: string; text: string; href: string }>;
-    promoBanners: Array<{ id: string; title: string; subtitle: string; code: string; href: string; imageUrl: string; priority: number; activeFrom: string; activeTo: string; ctaLabel: string }>;
-    adCards: Array<{ id: string; title: string; subtitle: string; href: string; imageUrl: string; badge: string; priority: number; activeFrom: string; activeTo: string; ctaLabel: string }>;
-    showcaseGroups: Array<{
-      id: string;
-      label: string;
-      mode: "subtype" | "price";
-      subtype?: string;
-      minPrice?: number;
-      maxPrice?: number;
-    }>;
-  };
-  ratingSummary: {
-    average: number;
-    count: number;
-  };
-  footer: {
-    socialLinks: Array<{ label: string; href: string }>;
-    textColor: string;
-  };
-}
-
-export interface FranchizeBySlugResult {
-  crew: FranchizeCrewVM;
-  items: CatalogItemVM[];
-}
-
-export interface FranchizeConfigInput {
-  slug: string;
-  brandName: string;
-  tagline: string;
-  logoUrl: string;
-  themeMode: string;
-  bgBase: string;
-  bgCard: string;
-  accentMain: string;
-  accentMainHover: string;
-  textPrimary: string;
-  textSecondary: string;
-  borderSoft: string;
-  lightBgBase: string;
-  lightBgCard: string;
-  lightAccentMain: string;
-  lightAccentMainHover: string;
-  lightTextPrimary: string;
-  lightTextSecondary: string;
-  lightBorderSoft: string;
-  phone: string;
-  email: string;
-  address: string;
-  telegram: string;
-  mapGps: string;
-  mapImageUrl: string;
-  mapBoundsTop: string;
-  mapBoundsBottom: string;
-  mapBoundsLeft: string;
-  mapBoundsRight: string;
-  socialLinksText: string;
-  menuLinksText: string;
-  categoryOrderText: string;
-  promoBannersText: string;
-  adCardsText: string;
-  allowPromo: boolean;
-  deliveryModesText: string;
-  paymentOptionsText: string;
-  defaultMode: string;
-  issuerName: string;
-  issuerRepresentative: string;
-  includedMileage: string;
-  overageRateRub: string;
-  bikeValueRub: string;
-  bikeValueWords: string;
-  lateReturnPenaltyRub: string;
-  returnAddress: string;
-  contractDefaultsJson: string;
-  docTemplatesJson: string;
-  advancedJson: string;
-}
-
-export interface FranchizeConfigState {
-  ok: boolean;
-  message: string;
-  errors?: Record<string, string[]>;
-  data?: FranchizeConfigInput;
-  canEdit?: boolean;
-}
-
-const defaultTheme = DEFAULT_FRANCHIZE_THEME;
-
-const franchizeConfigSchema = z.object({
-  slug: z.string().trim().min(2, "Slug is required"),
-  brandName: z.string().trim().min(2, "Brand name is required"),
-  tagline: z.string().trim().min(2, "Tagline is required"),
-  logoUrl: z.string().trim().optional(),
-  themeMode: z.string().trim().min(2, "Theme mode is required"),
-  bgBase: z.string().trim().min(4, "bgBase is required"),
-  bgCard: z.string().trim().min(4, "bgCard is required"),
-  accentMain: z.string().trim().min(4, "accentMain is required"),
-  accentMainHover: z.string().trim().min(4, "accentMainHover is required"),
-  textPrimary: z.string().trim().min(4, "textPrimary is required"),
-  textSecondary: z.string().trim().min(4, "textSecondary is required"),
-  borderSoft: z.string().trim().min(4, "borderSoft is required"),
-  lightBgBase: z.string().trim().min(4, "lightBgBase is required"),
-  lightBgCard: z.string().trim().min(4, "lightBgCard is required"),
-  lightAccentMain: z.string().trim().min(4, "lightAccentMain is required"),
-  lightAccentMainHover: z.string().trim().min(4, "lightAccentMainHover is required"),
-  lightTextPrimary: z.string().trim().min(4, "lightTextPrimary is required"),
-  lightTextSecondary: z.string().trim().min(4, "lightTextSecondary is required"),
-  lightBorderSoft: z.string().trim().min(4, "lightBorderSoft is required"),
-  phone: z.string().trim().default(""),
-  email: z.string().trim().default(""),
-  address: z.string().trim().default(""),
-  telegram: z.string().trim().default(""),
-  mapGps: z.string().trim().default(""),
-  mapImageUrl: z.string().trim().default(""),
-  mapBoundsTop: z.string().trim().default(String(DEFAULT_MAP_BOUNDS.top)),
-  mapBoundsBottom: z.string().trim().default(String(DEFAULT_MAP_BOUNDS.bottom)),
-  mapBoundsLeft: z.string().trim().default(String(DEFAULT_MAP_BOUNDS.left)),
-  mapBoundsRight: z.string().trim().default(String(DEFAULT_MAP_BOUNDS.right)),
-  socialLinksText: z.string().default(DEFAULT_SOCIAL_LINKS_TEXT),
-  menuLinksText: z.string().default(""),
-  categoryOrderText: z.string().default(""),
-  promoBannersText: z.string().default(""),
-  adCardsText: z.string().default(""),
-  allowPromo: z.coerce.boolean().default(true),
-  deliveryModesText: z.string().default(DEFAULT_DELIVERY_MODES_TEXT),
-  paymentOptionsText: z.string().default(DEFAULT_PAYMENT_OPTIONS_TEXT),
-  defaultMode: z.string().trim().default("pickup"),
-  issuerName: z.string().trim().default(""),
-  issuerRepresentative: z.string().trim().default(""),
-  includedMileage: z.string().trim().default(""),
-  overageRateRub: z.string().trim().default(""),
-  bikeValueRub: z.string().trim().default(""),
-  bikeValueWords: z.string().trim().default(""),
-  lateReturnPenaltyRub: z.string().trim().default(""),
-  returnAddress: z.string().trim().default(""),
-  contractDefaultsJson: z.string().default(""),
-  docTemplatesJson: z.string().default(""),
-  advancedJson: z.string().default(""),
-});
-
-const defaultFranchizeConfig: FranchizeConfigInput = {
-  slug: "",
-  brandName: DEFAULT_FRANCHIZE_BRAND.brandName,
-  tagline: DEFAULT_FRANCHIZE_BRAND.tagline,
-  logoUrl: "",
-  themeMode: defaultTheme.mode,
-  bgBase: defaultTheme.palette.bgBase,
-  bgCard: defaultTheme.palette.bgCard,
-  accentMain: defaultTheme.palette.accentMain,
-  accentMainHover: defaultTheme.palette.accentMainHover,
-  textPrimary: defaultTheme.palette.textPrimary,
-  textSecondary: defaultTheme.palette.textSecondary,
-  borderSoft: defaultTheme.palette.borderSoft,
-  lightBgBase: DEFAULT_LIGHT_THEME_PALETTE.bgBase,
-  lightBgCard: DEFAULT_LIGHT_THEME_PALETTE.bgCard,
-  lightAccentMain: DEFAULT_LIGHT_THEME_PALETTE.accentMain,
-  lightAccentMainHover: DEFAULT_LIGHT_THEME_PALETTE.accentMainHover,
-  lightTextPrimary: DEFAULT_LIGHT_THEME_PALETTE.textPrimary,
-  lightTextSecondary: DEFAULT_LIGHT_THEME_PALETTE.textSecondary,
-  lightBorderSoft: DEFAULT_LIGHT_THEME_PALETTE.borderSoft,
-  phone: "",
-  email: "",
-  address: "",
-  telegram: "",
-  mapGps: DEFAULT_MAP_GPS,
-  mapImageUrl: DEFAULT_MAP_IMAGE_URL,
-  mapBoundsTop: String(DEFAULT_MAP_BOUNDS.top),
-  mapBoundsBottom: String(DEFAULT_MAP_BOUNDS.bottom),
-  mapBoundsLeft: String(DEFAULT_MAP_BOUNDS.left),
-  mapBoundsRight: String(DEFAULT_MAP_BOUNDS.right),
-  socialLinksText: DEFAULT_SOCIAL_LINKS_TEXT,
-  menuLinksText: DEFAULT_MENU_LINK_TEMPLATES.map((link) => `${link.label}|${link.href}`).join("\n"),
-  categoryOrderText: DEFAULT_CATEGORY_ORDER,
-  promoBannersText: DEFAULT_PROMO_BANNERS_TEXT,
-  adCardsText: DEFAULT_AD_CARDS_TEXT,
-  allowPromo: true,
-  deliveryModesText: DEFAULT_DELIVERY_MODES_TEXT,
-  paymentOptionsText: DEFAULT_PAYMENT_OPTIONS_TEXT,
-  defaultMode: "pickup",
-  issuerName: "",
-  issuerRepresentative: "",
-  includedMileage: DEFAULT_CONTRACT_PREFILL.includedMileage,
-  overageRateRub: DEFAULT_CONTRACT_PREFILL.overageRateRub,
-  bikeValueRub: DEFAULT_CONTRACT_PREFILL.bikeValueRub,
-  bikeValueWords: DEFAULT_CONTRACT_PREFILL.bikeValueWords,
-  lateReturnPenaltyRub: DEFAULT_CONTRACT_PREFILL.lateReturnPenaltyRub,
-  returnAddress: DEFAULT_CONTRACT_PREFILL.returnAddress,
-  contractDefaultsJson: "",
-  docTemplatesJson: "",
-  advancedJson: "",
-};
-
-function readPath<T>(obj: unknown, path: string[], fallback: T): T {
-  let current: unknown = obj;
-  for (const key of path) {
-    if (!current || typeof current !== "object" || !(key in current)) {
-      return fallback;
-    }
-    current = (current as UnknownRecord)[key];
-  }
-  return (current as T) ?? fallback;
-}
-
-
-function toRentalReviewVM(row: UnknownRecord): RentalReviewVM {
-  return {
-    id: String(row.id ?? ""),
-    rentalId: String(row.rental_id ?? ""),
-    userId: String(row.user_id ?? ""),
-    bikeId: String(row.bike_id ?? ""),
-    crewId: String(row.crew_id ?? ""),
-    rating: Math.min(5, Math.max(1, Number(row.rating ?? 0) || 1)),
-    text: String(row.text ?? "").trim(),
-    createdAt: String(row.created_at ?? ""),
-    hiddenAt: typeof row.hidden_at === "string" ? row.hidden_at : null,
-  };
-}
-
-function buildReviewSummary(reviews: RentalReviewVM[]): RentalReviewSummaryVM {
-  const visible = reviews.filter((review) => !review.hiddenAt);
-  const count = visible.length;
-  const average = count > 0
-    ? Math.round((visible.reduce((sum, review) => sum + review.rating, 0) / count) * 10) / 10
-    : 0;
-  return {
-    average,
-    count,
-    latest: visible[0],
-    reviews: visible,
-  };
-}
-
-
-function groupReviewsByBike(rows: UnknownRecord[]): Map<string, RentalReviewVM[]> {
-  const grouped = new Map<string, RentalReviewVM[]>();
-  rows
-    .map(toRentalReviewVM)
-    .filter((review) => review.bikeId && !review.hiddenAt)
-    .forEach((review) => {
-      const list = grouped.get(review.bikeId) ?? [];
-      list.push(review);
-      grouped.set(review.bikeId, list);
-    });
-
-  grouped.forEach((list) => {
-    list.sort((a, b) => Date.parse(b.createdAt || "") - Date.parse(a.createdAt || ""));
-  });
-
-  return grouped;
-}
-
-function buildCrewRatingSummary(rows: UnknownRecord[]) {
-  return buildReviewSummary(rows.map(toRentalReviewVM).filter((review) => !review.hiddenAt));
-}
-
-/*const fallbackMenuLinks = (slug: string) => [
-  { label: "Каталог", href: `/franchize/${slug}` },
-  { label: "Электроэндуро", href: `/franchize/${slug}/electro-enduro` },
-  { label: "Конфигуратор", href: `/franchize/${slug}/configurator` },
-  { label: "Map Riders", href: `/franchize/${slug}/map-riders` },
-  { label: "О нас", href: `/franchize/${slug}/about` },
-  { label: "Контакты", href: `/franchize/${slug}/contacts` },
-  { label: "Корзина", href: `/franchize/${slug}/cart` },
-];*/
-function readArrayPath<T>(obj: unknown, path: string[], fallback: T[] = []): T[] {
-  const value = readPath<unknown>(obj, path, fallback);
-  return Array.isArray(value) ? (value as T[]) : fallback;
-}
-
-const fallbackMenuLinks = (slug: string) => DEFAULT_MENU_LINK_TEMPLATES.map((link) => ({
-  label: link.label,
-  href: withSlug(link.href, slug),
-}));
-
-function normalizeCrewSlug(value: string): string {
-  return String(value || "")
-    .trim()
-    .toLowerCase()
-    .replace(/%20/g, " ")
-    .replace(/\s+/g, "-")
-    .replace(/_+/g, "-")
-    .replace(/-+/g, "-");
-}
-
-const withSlug = (href: string, slug: string) => {
-  if (!href) {
-    return href;
-  }
-
-  if (href.includes("{slug}")) {
-    return href.replaceAll("{slug}", slug);
-  }
-
-  switch (href) {
-    case "/franchize/about":
-      return `/franchize/${slug}/about`;
-    case "/franchize/contacts":
-      return `/franchize/${slug}/contacts`;
-    case "/franchize/cart":
-      return `/franchize/${slug}/cart`;
-    case "/franchize/rentals":
-      return `/franchize/${slug}/rentals`;
-    default:
-      return href;
-  }
-};
-
-function parseSocialLinks(lines: string): Array<{ label: string; href: string }> {
-  return lines
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line) => {
-      const [label, href] = line.split("|").map((value) => value.trim());
-      return { label: label || "Social", href: href || DEFAULT_TELEGRAM_BOT_URL };
-    });
-}
-
-function extractFooterSocialLinks(franchize: UnknownRecord, fallbackTelegram: string) {
-  const explicit = readArrayPath<UnknownRecord>(franchize, ["footer", "socialLinks"]);
-  const fromExplicit = explicit
-    .map((item) => ({ label: readPath(item, ["label"], ""), href: readPath(item, ["href"], "") }))
-    .filter((item) => item.label && item.href);
-
-  if (fromExplicit.length > 0) return fromExplicit;
-
-  const columns = readArrayPath<UnknownRecord>(franchize, ["footer", "columns"]);
-  const fromColumns = columns.flatMap((column) => {
-    const items = readArrayPath<UnknownRecord>(column, ["items"]);
-    return items
-      .map((item) => ({
-        label: readPath(item, ["label"], readPath(item, ["value"], "")),
-        href: readPath(item, ["href"], ""),
-      }))
-      .filter((entry) => entry.label && entry.href);
-  });
-
-  if (fromColumns.length > 0) return fromColumns;
-
-  if (fallbackTelegram) {
-    return [{ label: fallbackTelegram, href: `https://t.me/${fallbackTelegram.replace("@", "")}` }];
-  }
-
-  return [{ label: "Telegram", href: DEFAULT_TELEGRAM_BOT_URL }];
-}
-
-const emptyCrew = (slug: string): FranchizeCrewVM => ({
-  id: "",
-  slug,
-  name: "Crew not found",
-  description: "No franchise crew found for this slug yet.",
-  logoUrl: "",
-  hqLocation: "",
-  isFound: false,
-  theme: defaultTheme,
-  header: {
-    brandName: "Franchize",
-    tagline: "Hydration pending",
-    logoUrl: "",
-    menuLinks: fallbackMenuLinks(slug),
-  },
-  contacts: {
-    phone: "",
-    email: "",
-    address: "",
-    telegram: "",
-    workingHours: "",
-    map: {
-      gps: "",
-      publicTransport: "",
-      carDirections: "",
-      imageUrl: "",
-      bounds: {
-        top: DEFAULT_MAP_BOUNDS.top,
-        bottom: DEFAULT_MAP_BOUNDS.bottom,
-        left: DEFAULT_MAP_BOUNDS.left,
-        right: DEFAULT_MAP_BOUNDS.right,
-      },
-    },
-  },
-  catalog: {
-    categories: [],
-    quickLinks: [],
-    tickerItems: [],
-    promoBanners: [],
-    adCards: [],
-    showcaseGroups: [],
-  },
-  ratingSummary: { average: 0, count: 0 },
-  footer: {
-    socialLinks: [],
-    textColor: "#16130A",
-  },
-});
-
-export async function getFranchizeBySlug(slug: string): Promise<FranchizeBySlugResult> {
-  const safeSlug = normalizeCrewSlug(slug ?? "");
-  if (!safeSlug) {
-    return { crew: emptyCrew("unknown"), items: [] };
-  }
-
-  try {
-    const { data: crew, error: crewError } = await supabaseAdmin
-      .from("crews")
-      .select("*")
-      .eq("slug", safeSlug)
-      .maybeSingle();
-
-    if (crewError || !crew) {
-      if (crewError) {
-        logger.warn("[franchize] failed to load crew by slug", { safeSlug, crewError: crewError.message });
-      }
-      return { crew: emptyCrew(safeSlug), items: [] };
-    }
-
-    const catalogTypes = ["bike", "accessories", "gear", "wbitem"];
-    const { data: cars, error: carsError } = await supabaseAdmin
-      .from("cars")
-      .select("id, make, model, description, image_url, daily_price, type, specs, availability_rules")
-      .eq("crew_id", crew.id)
-      .in("type", catalogTypes);
-
-    if (carsError) {
-      logger.warn("[franchize] failed to load crew cars", { safeSlug, carsError: carsError.message });
-    }
-
-    const vehicleIds = (cars ?? []).map((car) => car.id).filter(Boolean);
-    const { data: activeRentals, error: rentalsError } = vehicleIds.length
-      ? await supabaseAdmin
-          .from("rentals")
-          .select("vehicle_id, status, agreed_start_date, agreed_end_date, requested_end_date")
-          .in("vehicle_id", vehicleIds)
-          .in("status", ["pending_confirmation", "confirmed", "active"])
-      : { data: [], error: null };
-
-    if (rentalsError) {
-      logger.warn("[franchize] failed to load rentals availability", { safeSlug, rentalsError: rentalsError.message });
-    }
-
-    const availabilityByVehicle = new Map<string, { status: "available" | "busy"; label: string }>();
-    const nowTs = Date.now();
-    for (const rental of (activeRentals ?? []) as RentalAvailabilityRow[]) {
-      const vehicleId = typeof rental.vehicle_id === "string" ? rental.vehicle_id : "";
-      if (!vehicleId) continue;
-
-      const startTs = rental.agreed_start_date ? Date.parse(rental.agreed_start_date) : Number.NaN;
-      const endTs = rental.agreed_end_date
-        ? Date.parse(rental.agreed_end_date)
-        : rental.requested_end_date
-          ? Date.parse(rental.requested_end_date)
-          : Number.NaN;
-      const rentalStatus = typeof rental.status === "string" ? rental.status : "pending_confirmation";
-
-      const isActiveNow = rentalStatus === "active" || (!Number.isNaN(startTs) && startTs <= nowTs && (Number.isNaN(endTs) || endTs >= nowTs));
-      const isUpcomingBooking = ["pending_confirmation", "confirmed"].includes(rentalStatus) && (Number.isNaN(endTs) || endTs >= nowTs);
-      const isBusy = isActiveNow || isUpcomingBooking;
-      if (!isBusy) continue;
-
-      const label = isActiveNow
-        ? (Number.isNaN(endTs) ? "В аренде сейчас" : `В аренде до ${new Date(endTs).toLocaleDateString("ru-RU")}`)
-        : (Number.isNaN(startTs) ? "Забронирован" : `Забронирован с ${new Date(startTs).toLocaleDateString("ru-RU")}`);
-      availabilityByVehicle.set(vehicleId, { status: "busy", label });
-    }
-
-    const { data: reviewRows, error: reviewsError } = vehicleIds.length
-      ? await supabaseAdmin
-          .from("rental_reviews")
-          .select("id, rental_id, user_id, bike_id, crew_id, rating, text, created_at, hidden_at")
-          .in("bike_id", vehicleIds)
-          .is("hidden_at", null)
-          .order("created_at", { ascending: false })
-      : { data: [], error: null };
-
-    if (reviewsError) {
-      logger.warn("[franchize] failed to load rental reviews", { safeSlug, reviewsError: reviewsError.message });
-    }
-
-    const reviewRowsRaw = ((reviewRows ?? []) as UnknownRecord[]);
-    const reviewsByBike = groupReviewsByBike(reviewRowsRaw);
-    const crewRatingSummary = buildCrewRatingSummary(reviewRowsRaw);
-
-    const metadata = ((crew as UnknownRecord).metadata ?? {}) as UnknownRecord;
-    const franchize = (metadata.franchize ?? metadata) as UnknownRecord;
-
-    const resolvedTheme = resolveFranchizeTheme(franchize);
-    const menuLinksRaw = readArrayPath<UnknownRecord>(franchize, ["header", "menuLinks"], fallbackMenuLinks(safeSlug)).map((link) => ({
-      label: readPath(link, ["label"], "Link"),
-      href: withSlug(readPath(link, ["href"], `/franchize/${crew.slug ?? safeSlug}`), crew.slug ?? safeSlug),
-    }));
-    const menuLinks = menuLinksRaw.some((link) => link.href === `/franchize/${crew.slug ?? safeSlug}/map-riders`)
-      ? menuLinksRaw
-      : [
-          ...menuLinksRaw.slice(0, 1),
-          { label: "Map Riders", href: `/franchize/${crew.slug ?? safeSlug}/map-riders` },
-          ...menuLinksRaw.slice(1),
-        ];
-
-    const metadataShowcaseGroups = readArrayPath<UnknownRecord>(franchize, ["catalog", "showcaseGroups"]);
-    const showcaseGroups = (metadataShowcaseGroups.length > 0
-      ? metadataShowcaseGroups.map((group, index) => ({
-          id: readPath(group, ["id"], `showcase-${index + 1}`),
-          label: readPath(group, ["label"], readPath(group, ["title"], `Showcase ${index + 1}`)),
-          mode: readPath(group, ["mode"], "subtype") as "subtype" | "price",
-          subtype: readPath(group, ["subtype"], ""),
-          minPrice: Number(readPath(group, ["minPrice"], 0)),
-          maxPrice: Number(readPath(group, ["maxPrice"], 0)),
-        }))
-      : DEFAULT_SHOWCASE_GROUPS)
-      .filter((group) => group.label.trim().length > 0);
-
-    const hydratedCrew: FranchizeCrewVM = {
-      id: crew.id,
-      slug: crew.slug ?? safeSlug,
-      name: crew.name ?? "Unnamed crew",
-      description: crew.description ?? "",
-      logoUrl: crew.logo_url ?? "",
-      hqLocation: crew.hq_location ?? "",
-      isFound: true,
-      theme: resolvedTheme,
-      header: {
-        brandName: readPath(franchize, ["branding", "name"], crew.name ?? "Franchize"),
-        tagline: readPath(franchize, ["branding", "tagline"], "Ride the vibe"),
-        logoUrl: readPath(franchize, ["branding", "logoUrl"], crew.logo_url ?? ""),
-        logoHref: readPath(franchize, ["header", "logoHref"], `/franchize/${crew.slug ?? safeSlug}`),
-        menuLinks,
-      },
-      contacts: {
-        phone: readPath(franchize, ["contacts", "phone"], readPath(franchize, ["footer", "phone"], "")),
-        email: readPath(franchize, ["contacts", "email"], readPath(franchize, ["footer", "email"], "")),
-        address: readPath(
-          franchize,
-          ["contacts", "address"],
-          readPath(franchize, ["footer", "address"], crew.hq_location ?? ""),
-        ),
-        telegram: readPath(franchize, ["contacts", "telegram"], ""),
-        workingHours: readPath(franchize, ["contacts", "workingHours"], ""),
-        map: {
-          gps: readPath(franchize, ["contacts", "map", "gps"], ""),
-          publicTransport: readPath(franchize, ["contacts", "map", "publicTransport"], ""),
-          carDirections: readPath(franchize, ["contacts", "map", "carDirections"], ""),
-          imageUrl: readPath(franchize, ["contacts", "map", "imageUrl"], ""),
-          bounds: {
-            top: Number(readPath(franchize, ["contacts", "map", "bounds", "top"], DEFAULT_MAP_BOUNDS.top)),
-            bottom: Number(readPath(franchize, ["contacts", "map", "bounds", "bottom"], DEFAULT_MAP_BOUNDS.bottom)),
-            left: Number(readPath(franchize, ["contacts", "map", "bounds", "left"], DEFAULT_MAP_BOUNDS.left)),
-            right: Number(readPath(franchize, ["contacts", "map", "bounds", "right"], DEFAULT_MAP_BOUNDS.right)),
-          },
-        },
-      },
-      catalog: {
-        categories: readArrayPath<string>(franchize, ["catalog", "groupOrder"]),
-        quickLinks: readArrayPath<string>(franchize, ["catalog", "quickLinks"]),
-        tickerItems: readArrayPath<unknown>(franchize, ["catalog", "tickerItems"]).map((item: unknown, index: number) => {
-          const tickerItem = (item ?? {}) as UnknownRecord;
-          const text = readPath(tickerItem, ["text"], readPath(tickerItem, ["title"], ""));
-
-          return {
-            id: readPath(tickerItem, ["id"], `ticker-${index + 1}`),
-            text,
-            href: readPath(tickerItem, ["href"], `/franchize/${crew.slug ?? safeSlug}`),
-          };
-        }),
-        promoBanners: readArrayPath<unknown>(franchize, ["catalog", "promoBanners"]).map((item: unknown, index: number) => {
-          const banner = (item ?? {}) as UnknownRecord;
-          return {
-            id: readPath(banner, ["id"], `promo-${index + 1}`),
-            title: readPath(banner, ["title"], ""),
-            subtitle: readPath(banner, ["subtitle"], ""),
-            code: readPath(banner, ["code"], ""),
-            href: readPath(banner, ["href"], `/franchize/${crew.slug ?? safeSlug}`),
-            imageUrl: readPath(banner, ["imageUrl"], ""),
-            priority: Number(readPath(banner, ["priority"], 50)),
-            activeFrom: readPath(banner, ["activeFrom"], ""),
-            activeTo: readPath(banner, ["activeTo"], ""),
-            ctaLabel: readPath(banner, ["ctaLabel"], "Открыть"),
-          };
-        }),
-        adCards: readArrayPath<unknown>(franchize, ["catalog", "adCards"]).map((item: unknown, index: number) => {
-          const ad = (item ?? {}) as UnknownRecord;
-          return {
-            id: readPath(ad, ["id"], `ad-${index + 1}`),
-            title: readPath(ad, ["title"], ""),
-            subtitle: readPath(ad, ["subtitle"], ""),
-            href: readPath(ad, ["href"], `/franchize/${crew.slug ?? safeSlug}`),
-            imageUrl: readPath(ad, ["imageUrl"], ""),
-            badge: readPath(ad, ["badge"], ""),
-            priority: Number(readPath(ad, ["priority"], 40)),
-            activeFrom: readPath(ad, ["activeFrom"], ""),
-            activeTo: readPath(ad, ["activeTo"], ""),
-            ctaLabel: readPath(ad, ["ctaLabel"], "Подробнее"),
-          };
-        }),
-        showcaseGroups,
-      },
-      ratingSummary: { average: crewRatingSummary.average, count: crewRatingSummary.count },
-      footer: {
-        socialLinks: extractFooterSocialLinks(franchize, readPath(franchize, ["contacts", "telegram"], "")),
-        textColor: readPath(franchize, ["footer", "textColor"], DEFAULT_FOOTER_TEXT_COLOR),
-      },
-    };
-
-    const items: CatalogItemVM[] = (cars ?? []).map((car) => {
-      const specs = (car.specs ?? {}) as UnknownRecord;
-      const subtype =
-        (typeof specs.subtype === "string" && specs.subtype.trim()) ||
-        (typeof specs.bike_subtype === "string" && specs.bike_subtype.trim()) ||
-        (typeof specs.segment === "string" && specs.segment.trim()) ||
-        (typeof specs.type === "string" && specs.type.trim().toLowerCase() !== "bike" && specs.type.trim()) ||
-        "Unsorted";
-
-      const mediaUrls = Array.from(new Set([
-        car.image_url ?? "",
-        ...(Array.isArray(specs.gallery) ? specs.gallery.filter((value): value is string => typeof value === "string") : []),
-        ...(Array.isArray(specs.images) ? specs.images.filter((value): value is string => typeof value === "string") : []),
-        ...(Array.isArray(specs.photos) ? specs.photos.filter((value): value is string => typeof value === "string") : []),
-        ...(Array.isArray(specs.image_urls) ? specs.image_urls.filter((value): value is string => typeof value === "string") : []),
-      ].map((value) => value.trim()).filter(Boolean)));
-
-      const availabilityRulesRaw = (
-        car.availability_rules && typeof car.availability_rules === "object" && !Array.isArray(car.availability_rules)
-          ? (car.availability_rules as Record<string, unknown>)
-          : {}
-      );
-      const rulesType = typeof availabilityRulesRaw.type === "string" ? availabilityRulesRaw.type : "";
-      const manualStatus = typeof availabilityRulesRaw.manual_status === "string"
-        ? availabilityRulesRaw.manual_status
-        : typeof availabilityRulesRaw.manualStatus === "string"
-          ? availabilityRulesRaw.manualStatus
-          : "";
-      const isWeekend = [6, 7].includes(new Date().getDay() === 0 ? 7 : new Date().getDay());
-      const blockedByRules =
-        manualStatus === "busy" ||
-        manualStatus === "unavailable" ||
-        (rulesType === "weekends_only" && !isWeekend);
-
-      const fallbackStatus: "available" | "busy" = blockedByRules ? "busy" : "available";
-      const fallbackLabel =
-        manualStatus === "busy"
-          ? "Временно занят"
-          : manualStatus === "unavailable"
-            ? "Временно недоступен"
-            : rulesType === "weekends_only" && !isWeekend
-              ? "Доступен только по выходным"
-              : "Свободен сегодня";
-
-      return {
-        id: car.id,
-        title: `${car.make} ${car.model}`.trim(),
-        subtitle: (typeof specs.subtitle === "string" ? specs.subtitle : "Ready to ride") as string,
-        description: car.description ?? (typeof specs.description === "string" ? specs.description : "Подробности откроются в карточке"),
-        imageUrl: mediaUrls[0] ?? "",
-        mediaUrls,
-        pricePerDay: car.daily_price ?? 0,
-        rentPriceLabel:
-          (typeof readPath(specs, ["rent_price_label"], "") === "string" && String(readPath(specs, ["rent_price_label"], "")).trim().length > 0)
-            ? String(readPath(specs, ["rent_price_label"], "")).trim()
-            : `${Number(car.daily_price ?? 0).toLocaleString("ru-RU")} ₽ / день`,
-        category: subtype,
-        availabilityStatus: availabilityByVehicle.get(car.id)?.status ?? fallbackStatus,
-        availabilityLabel: availabilityByVehicle.get(car.id)?.label ?? fallbackLabel,
-        isHot:
-          Number(car.daily_price ?? 0) >= 7000 ||
-          Boolean(readPath(specs, ["is_hot"], false)) ||
-          Boolean(readPath(specs, ["hot"], false)) ||
-          /hot|🔥/i.test(String(readPath(specs, ["badge"], ""))),
-        saleAvailable:
-          readPath(specs, ["sale"], 0) === 1 ||
-          readPath(specs, ["sale"], false) === true ||
-          String(readPath(specs, ["sale"], "")).toLowerCase() === "1" ||
-          String(readPath(specs, ["sale"], "")).toLowerCase() === "true",
-        salePrice:
-          Number(readPath(specs, ["sale_price"], 0)) > 0
-            ? Number(readPath(specs, ["sale_price"], 0))
-            : Number(readPath(specs, ["purchase_price"], 0)) > 0
-              ? Number(readPath(specs, ["purchase_price"], 0))
-              : null,
-        rawSpecs: specs,
-        reviewSummary: buildReviewSummary(reviewsByBike.get(car.id) ?? []),
-        specs: Object.entries(specs)
-          .filter(([, value]) => typeof value === "string" || typeof value === "number")
-          .filter(([key]) => !["subtitle", "description", "segment", "subtype", "bike_subtype", "type"].includes(key))
-          .slice(0, 4)
-          .map(([key, value]) => ({ label: key.replace(/_/g, " "), value: String(value) })),
-      };
-    });
-
-    return {
-      crew: {
-        ...hydratedCrew,
-        catalog: {
-          categories:
-            hydratedCrew.catalog.categories.length > 0
-              ? normalizeCatalogOrder(hydratedCrew.catalog.categories)
-              : normalizeCatalogOrder(items.map((item) => item.category)),
-          quickLinks: hydratedCrew.catalog.quickLinks,
-          tickerItems: hydratedCrew.catalog.tickerItems.filter((item) => item.text.trim().length > 0),
-          promoBanners: hydratedCrew.catalog.promoBanners.filter((item) => item.title.trim().length > 0),
-          adCards: hydratedCrew.catalog.adCards.filter((item) => item.title.trim().length > 0),
-          showcaseGroups: hydratedCrew.catalog.showcaseGroups,
-        },
-        footer: hydratedCrew.footer,
-      },
-      items,
-    };
-  } catch (error) {
-    logger.error("[franchize] unexpected getFranchizeBySlug failure", error);
-    return { crew: emptyCrew(safeSlug), items: [] };
-  }
-}
-
-export async function markCrewBikesAvailable(slug: string) {
-  const safeSlug = slug?.trim().toLowerCase().replace(/\s+/g, "-");
-  if (!safeSlug) {
-    return { success: false, error: "slug is required" };
-  }
-
-  const { data: crew, error: crewError } = await supabaseAdmin
-    .from("crews")
-    .select("id")
-    .eq("slug", safeSlug)
-    .maybeSingle();
-
-  if (crewError || !crew) {
-    return { success: false, error: crewError?.message || "crew not found" };
-  }
-
-  const { data: cars, error: carsError } = await supabaseAdmin
-    .from("cars")
-    .select("id, availability_rules")
-    .eq("crew_id", crew.id)
-    .in("type", ["bike", "ebike"]);
-
-  if (carsError) {
-    return { success: false, error: carsError.message };
-  }
-
-  for (const car of cars ?? []) {
-    const currentRules = (car.availability_rules && typeof car.availability_rules === "object" && !Array.isArray(car.availability_rules))
-      ? (car.availability_rules as Record<string, unknown>)
-      : {};
-
-    const nextRules = {
-      ...currentRules,
-      manual_status: "available",
-      initialized_by: "markCrewBikesAvailable",
-      initialized_at: new Date().toISOString(),
-    };
-
-    const { error } = await supabaseAdmin
-      .from("cars")
-      .update({ availability_rules: nextRules })
-      .eq("id", car.id);
-
-    if (error) {
-      return { success: false, error: error.message };
-    }
-  }
-
-  return { success: true };
-}
-
-function splitCsv(text: string): string[] {
-  return text
-    .split(",")
-    .map((value) => value.trim())
-    .filter(Boolean);
-}
-
-function normalizeCampaignHref(href: string, slug: string): string {
-  const clean = href.trim();
-  if (!clean) {
-    return `/franchize/${slug}#catalog-sections`;
-  }
-
-  if (clean.startsWith("http://") || clean.startsWith("https://") || clean.startsWith("mailto:") || clean.startsWith("tel:")) {
-    return clean;
-  }
-
-  return withSlug(clean, slug);
-}
-
-function trimCampaignTitle(title: string, fallback: string): string {
-  const normalized = title.replace(/\s+/g, " ").trim();
-  if (!normalized) {
-    return fallback;
-  }
-
-  return normalized.length > 72 ? `${normalized.slice(0, 69)}...` : normalized;
-}
-
-function parsePromoBanners(lines: string, slug: string): Array<{ id: string; title: string; subtitle: string; code: string; href: string; imageUrl: string; activeFrom: string; activeTo: string; priority: number; ctaLabel: string }> {
-  return lines
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line, index) => {
-      const [id, title, subtitle, code, href, imageUrl, activeFrom, activeTo, priority, ctaLabel] = line.split("|").map((value) => value.trim());
-      return {
-        id: id || `promo-${index + 1}`,
-        title: trimCampaignTitle(title || "", `Promo ${index + 1}`),
-        subtitle: subtitle || "",
-        code: code || "",
-        href: normalizeCampaignHref(href || "", slug),
-        imageUrl: imageUrl || "",
-        activeFrom: activeFrom || "",
-        activeTo: activeTo || "",
-        priority: Number.isFinite(Number(priority)) ? Number(priority) : 50,
-        ctaLabel: ctaLabel || "Открыть",
-      };
-    });
-}
-
-function parseAdCards(lines: string, slug: string): Array<{ id: string; title: string; subtitle: string; href: string; imageUrl: string; badge: string; activeFrom: string; activeTo: string; priority: number; ctaLabel: string }> {
-  return lines
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line, index) => {
-      const [id, title, subtitle, href, imageUrl, badge, activeFrom, activeTo, priority, ctaLabel] = line.split("|").map((value) => value.trim());
-      return {
-        id: id || `ad-${index + 1}`,
-        title: trimCampaignTitle(title || "", `Ad ${index + 1}`),
-        subtitle: subtitle || "",
-        href: normalizeCampaignHref(href || "", slug),
-        imageUrl: imageUrl || "",
-        badge: badge || "Ad",
-        activeFrom: activeFrom || "",
-        activeTo: activeTo || "",
-        priority: Number.isFinite(Number(priority)) ? Number(priority) : 40,
-        ctaLabel: ctaLabel || "Подробнее",
-      };
-    });
-}
-
-function parseMenuLinks(lines: string, slug: string): Array<{ label: string; href: string }> {
-  return lines
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean)
-    .map((line) => {
-      const [label, href] = line.split("|").map((value) => value.trim());
-      return {
-        label: label || "Link",
-        href: withSlug(href || `/franchize/${slug}`, slug),
-      };
-    });
-}
-
-function normalizeCatalogOrder(categories: string[]): string[] {
-  const unique = Array.from(new Set(categories.filter(Boolean)));
-  const regular = unique.filter((category) => !category.toLowerCase().includes("wbitem"));
-  const wbItems = unique.filter((category) => category.toLowerCase().includes("wbitem"));
-  return [...regular, ...wbItems];
-}
-
-async function resolveFranchizeEditorAccess(actorUserId: string | undefined, crew: { id: string; owner_id?: string | null }): Promise<boolean> {
-  if (!actorUserId) {
-    return false;
-  }
-
-  const { data: user } = await supabaseAdmin
-    .from("users")
-    .select("role, status")
-    .eq("user_id", actorUserId)
-    .maybeSingle();
-
-  const isAdmin = user?.status === "admin" || user?.role === "admin" || user?.role === "vprAdmin";
-  if (isAdmin) {
-    return true;
-  }
-
-  if (crew.owner_id && crew.owner_id === actorUserId) {
-    return true;
-  }
-
-  const { data: membership } = await supabaseAdmin
-    .from("crew_members")
-    .select("role, membership_status")
-    .eq("crew_id", crew.id)
-    .eq("user_id", actorUserId)
-    .maybeSingle();
-
-  return membership?.membership_status === "active" && membership.role === "owner";
-}
-
-async function toFranchizeConfigInput(crew: UnknownRecord, slug: string): Promise<FranchizeConfigInput> {
-  const metadata = (crew.metadata ?? {}) as UnknownRecord;
-  const franchize = (metadata.franchize ?? {}) as UnknownRecord;
-  const themePalette = resolvePaletteByMode(franchize);
-  const lightPalette = {
-    bgBase: readPath(franchize, ["theme", "palettes", "light", "bgBase"], readPath(franchize, ["theme", "palette", "light", "bgBase"], defaultFranchizeConfig.lightBgBase)),
-    bgCard: readPath(franchize, ["theme", "palettes", "light", "bgCard"], readPath(franchize, ["theme", "palette", "light", "bgCard"], defaultFranchizeConfig.lightBgCard)),
-    accentMain: readPath(franchize, ["theme", "palettes", "light", "accentMain"], readPath(franchize, ["theme", "palette", "light", "accentMain"], defaultFranchizeConfig.lightAccentMain)),
-    accentMainHover: readPath(franchize, ["theme", "palettes", "light", "accentMainHover"], readPath(franchize, ["theme", "palette", "light", "accentMainHover"], defaultFranchizeConfig.lightAccentMainHover)),
-    textPrimary: readPath(franchize, ["theme", "palettes", "light", "textPrimary"], readPath(franchize, ["theme", "palette", "light", "textPrimary"], defaultFranchizeConfig.lightTextPrimary)),
-    textSecondary: readPath(franchize, ["theme", "palettes", "light", "textSecondary"], readPath(franchize, ["theme", "palette", "light", "textSecondary"], defaultFranchizeConfig.lightTextSecondary)),
-    borderSoft: readPath(franchize, ["theme", "palettes", "light", "borderSoft"], readPath(franchize, ["theme", "palette", "light", "borderSoft"], defaultFranchizeConfig.lightBorderSoft)),
-  };
-  const menuLinks = readArrayPath<UnknownRecord>(franchize, ["header", "menuLinks"], fallbackMenuLinks(slug));
-
-  const crewSecrets = await getCrewSensitiveData(slug);
-  const contractDefaults = (crewSecrets.contractDefaults ?? {}) as UnknownRecord;
-  const defaults = (readPath(contractDefaults, ["defaults"], {}) ?? {}) as UnknownRecord;
-  const docTemplates = (crewSecrets.docTemplates ?? {}) as UnknownRecord;
-
-  return {
-    ...defaultFranchizeConfig,
-    slug,
-    brandName: readPath(franchize, ["branding", "name"], (crew.name as string) ?? defaultFranchizeConfig.brandName),
-    tagline: readPath(franchize, ["branding", "tagline"], defaultFranchizeConfig.tagline),
-    logoUrl: readPath(franchize, ["branding", "logoUrl"], (crew.logo_url as string) ?? ""),
-    themeMode: readPath(franchize, ["theme", "mode"], defaultTheme.mode),
-    bgBase: readPath(themePalette, ["bgBase"], defaultTheme.palette.bgBase),
-    bgCard: readPath(themePalette, ["bgCard"], defaultTheme.palette.bgCard),
-    accentMain: readPath(themePalette, ["accentMain"], defaultTheme.palette.accentMain),
-    accentMainHover: readPath(themePalette, ["accentMainHover"], defaultTheme.palette.accentMainHover),
-    textPrimary: readPath(themePalette, ["textPrimary"], defaultTheme.palette.textPrimary),
-    textSecondary: readPath(themePalette, ["textSecondary"], defaultTheme.palette.textSecondary),
-    borderSoft: readPath(themePalette, ["borderSoft"], defaultTheme.palette.borderSoft),
-    lightBgBase: lightPalette.bgBase,
-    lightBgCard: lightPalette.bgCard,
-    lightAccentMain: lightPalette.accentMain,
-    lightAccentMainHover: lightPalette.accentMainHover,
-    lightTextPrimary: lightPalette.textPrimary,
-    lightTextSecondary: lightPalette.textSecondary,
-    lightBorderSoft: lightPalette.borderSoft,
-    phone: readPath(franchize, ["contacts", "phone"], readPath(franchize, ["footer", "phone"], "")),
-    email: readPath(franchize, ["contacts", "email"], readPath(franchize, ["footer", "email"], "")),
-    address: readPath(franchize, ["contacts", "address"], readPath(franchize, ["footer", "address"], "")),
-    telegram: readPath(franchize, ["contacts", "telegram"], ""),
-    mapGps: readPath(franchize, ["contacts", "map", "gps"], ""),
-    mapImageUrl: readPath(franchize, ["contacts", "map", "imageUrl"], ""),
-    mapBoundsTop: String(readPath(franchize, ["contacts", "map", "bounds", "top"], DEFAULT_MAP_BOUNDS.top)),
-    mapBoundsBottom: String(readPath(franchize, ["contacts", "map", "bounds", "bottom"], DEFAULT_MAP_BOUNDS.bottom)),
-    mapBoundsLeft: String(readPath(franchize, ["contacts", "map", "bounds", "left"], DEFAULT_MAP_BOUNDS.left)),
-    mapBoundsRight: String(readPath(franchize, ["contacts", "map", "bounds", "right"], DEFAULT_MAP_BOUNDS.right)),
-    socialLinksText: extractFooterSocialLinks(franchize, readPath(franchize, ["contacts", "telegram"], ""))
-      .map((entry) => `${entry.label}|${entry.href}`)
-      .join("\n"),
-    menuLinksText: menuLinks
-      .map((entry) => `${readPath(entry, ["label"], "Link")}|${readPath(entry, ["href"], `/franchize/${slug}`)}`)
-      .join("\n"),
-    categoryOrderText: readArrayPath<string>(franchize, ["catalog", "groupOrder"]).join(", "),
-    promoBannersText: readArrayPath<unknown>(franchize, ["catalog", "promoBanners"])
-      .map((entry: unknown, index: number) => {
-        const row = (entry ?? {}) as UnknownRecord;
-        return [
-          readPath(row, ["id"], `promo-${index + 1}`),
-          readPath(row, ["title"], ""),
-          readPath(row, ["subtitle"], ""),
-          readPath(row, ["code"], ""),
-          readPath(row, ["href"], ""),
-          readPath(row, ["imageUrl"], ""),
-          readPath(row, ["activeFrom"], ""),
-          readPath(row, ["activeTo"], ""),
-          String(readPath(row, ["priority"], 50)),
-          readPath(row, ["ctaLabel"], ""),
-        ].join("|");
-      })
-      .join("\n"),
-    adCardsText: readArrayPath<unknown>(franchize, ["catalog", "adCards"])
-      .map((entry: unknown, index: number) => {
-        const row = (entry ?? {}) as UnknownRecord;
-        return [
-          readPath(row, ["id"], `ad-${index + 1}`),
-          readPath(row, ["title"], ""),
-          readPath(row, ["subtitle"], ""),
-          readPath(row, ["href"], ""),
-          readPath(row, ["imageUrl"], ""),
-          readPath(row, ["badge"], ""),
-          readPath(row, ["activeFrom"], ""),
-          readPath(row, ["activeTo"], ""),
-          String(readPath(row, ["priority"], 40)),
-          readPath(row, ["ctaLabel"], ""),
-        ].join("|");
-      })
-      .join("\n"),
-    allowPromo: readPath(franchize, ["order", "allowPromo"], true),
-    deliveryModesText: readArrayPath<string>(franchize, ["order", "deliveryModes"], DEFAULT_DELIVERY_MODES_TEXT.split(", ")).join(", "),
-    paymentOptionsText: readArrayPath<string>(franchize, ["order", "paymentOptions"], DEFAULT_PAYMENT_OPTIONS_TEXT.split(", ")).join(", "),
-    defaultMode: readPath(franchize, ["order", "defaultMode"], "pickup"),
-    issuerName: readPath(defaults, ["issuerName"], ""),
-    issuerRepresentative: readPath(defaults, ["issuer_representative"], ""),
-    includedMileage: String(readPath(defaults, ["included_mileage"], defaultFranchizeConfig.includedMileage)),
-    overageRateRub: String(readPath(defaults, ["overage_rate"], defaultFranchizeConfig.overageRateRub)),
-    bikeValueRub: String(readPath(defaults, ["bike_value_rub"], defaultFranchizeConfig.bikeValueRub)),
-    bikeValueWords: readPath(defaults, ["bike_value_words"], defaultFranchizeConfig.bikeValueWords),
-    lateReturnPenaltyRub: String(readPath(defaults, ["late_return_penalty_rub"], defaultFranchizeConfig.lateReturnPenaltyRub)),
-    returnAddress: readPath(defaults, ["return_address"], defaultFranchizeConfig.returnAddress),
-    contractDefaultsJson: JSON.stringify(contractDefaults, null, 2),
-    docTemplatesJson: JSON.stringify(docTemplates, null, 2),
-    advancedJson: JSON.stringify(franchize, null, 2),
-  };
-}
-
-export async function loadFranchizeConfigBySlug(slug: string, actorUserId?: string): Promise<FranchizeConfigState> {
-  const safeSlug = normalizeCrewSlug(String(slug ?? ""));
-  if (!safeSlug) {
-    return { ok: false, message: "Сначала укажите slug экипажа." };
-  }
-
-  const { data: crew, error } = await supabaseAdmin.from("crews").select("id, slug, name, logo_url, metadata, owner_id").eq("slug", safeSlug).maybeSingle();
-  if (error || !crew) {
-    return { ok: false, message: "Экипаж с таким slug не найден." };
-  }
-
-  const canEdit = await resolveFranchizeEditorAccess(actorUserId, { id: crew.id, owner_id: crew.owner_id });
-
-  return {
-    ok: true,
-    message: canEdit
-      ? `Конфиг для ${safeSlug} загружен.`
-      : `Конфиг для ${safeSlug} загружен в read-only режиме.`,
-    data: await toFranchizeConfigInput(crew as UnknownRecord, safeSlug),
-    canEdit,
-  };
-}
-
-export async function saveFranchizeConfig(input: FranchizeConfigInput, actorUserId?: string): Promise<FranchizeConfigState> {
-  const parsed = franchizeConfigSchema.safeParse(input);
-  if (!parsed.success) {
-    return {
-      ok: false,
-      message: "Проверка не пройдена: проверьте обязательные поля формы.",
-      errors: parsed.error.flatten().fieldErrors,
-      data: input,
-    };
-  }
-
-  const payload = parsed.data;
-  const normalizedSlug = normalizeCrewSlug(payload.slug);
-  payload.slug = normalizedSlug;
-
-  const { data: crew, error: crewError } = await supabaseAdmin.from("crews").select("id, slug, metadata, owner_id").eq("slug", normalizedSlug).maybeSingle();
-  if (crewError || !crew) {
-    return { ok: false, message: "Slug не найден: сохранение не выполнено.", data: payload };
-  }
-
-  const canEdit = await resolveFranchizeEditorAccess(actorUserId, { id: crew.id, owner_id: crew.owner_id });
-  if (!canEdit) {
-    return { ok: false, message: "Недостаточно прав: сохранять может только owner экипажа или all-admin.", data: payload, canEdit: false };
-  }
-
-  let advancedOverrides: UnknownRecord = {};
-  if (payload.advancedJson.trim()) {
-    try {
-      const parsedJson = JSON.parse(payload.advancedJson) as unknown;
-      if (parsedJson && typeof parsedJson === "object") {
-        advancedOverrides = parsedJson as UnknownRecord;
-      }
-    } catch {
-      return { ok: false, message: "Advanced JSON должен быть валидным JSON-объектом.", data: payload };
-    }
-  }
-
-  let contractDefaultsOverrides: UnknownRecord = {};
-  if (payload.contractDefaultsJson.trim()) {
-    try {
-      const parsedJson = JSON.parse(payload.contractDefaultsJson) as unknown;
-      if (parsedJson && typeof parsedJson === "object" && !Array.isArray(parsedJson)) {
-        contractDefaultsOverrides = parsedJson as UnknownRecord;
-      }
-    } catch {
-      return { ok: false, message: "Contract defaults JSON должен быть валидным JSON-объектом.", data: payload };
-    }
-  }
-
-  let docTemplatesOverrides: UnknownRecord = {};
-  if (payload.docTemplatesJson.trim()) {
-    try {
-      const parsedJson = JSON.parse(payload.docTemplatesJson) as unknown;
-      if (parsedJson && typeof parsedJson === "object" && !Array.isArray(parsedJson)) {
-        docTemplatesOverrides = parsedJson as UnknownRecord;
-      }
-    } catch {
-      return { ok: false, message: "Doc templates JSON должен быть валидным JSON-объектом.", data: payload };
-    }
-  }
-
-  const baseMetadata = ((crew as UnknownRecord).metadata ?? {}) as UnknownRecord;
-  const existingFranchize = (readPath(baseMetadata, ["franchize"], {}) ?? {}) as UnknownRecord;
-  const menuLinks = parseMenuLinks(payload.menuLinksText, payload.slug);
-
-  const sourceFranchize: UnknownRecord = {
-    ...existingFranchize,
-    ...advancedOverrides,
-  };
-
-  const franchizeMetadata: UnknownRecord = {
-    ...sourceFranchize,
-    version: readPath(sourceFranchize, ["version"], "2026-02-19-editor-v2"),
-    enabled: readPath(sourceFranchize, ["enabled"], true),
-    slug: payload.slug,
-    branding: {
-      ...(readPath(sourceFranchize, ["branding"], {}) as UnknownRecord),
-      name: payload.brandName,
-      tagline: payload.tagline,
-      logoUrl: payload.logoUrl,
-    },
-    theme: {
-      ...(readPath(sourceFranchize, ["theme"], {}) as UnknownRecord),
-      mode: payload.themeMode,
-      palette: {
-        ...(readPath(sourceFranchize, ["theme", "palette"], {}) as UnknownRecord),
-        bgBase: payload.themeMode.toLowerCase().includes("light") ? payload.lightBgBase : payload.bgBase,
-        bgCard: payload.themeMode.toLowerCase().includes("light") ? payload.lightBgCard : payload.bgCard,
-        accentMain: payload.themeMode.toLowerCase().includes("light") ? payload.lightAccentMain : payload.accentMain,
-        accentMainHover: payload.themeMode.toLowerCase().includes("light") ? payload.lightAccentMainHover : payload.accentMainHover,
-        textPrimary: payload.themeMode.toLowerCase().includes("light") ? payload.lightTextPrimary : payload.textPrimary,
-        textSecondary: payload.themeMode.toLowerCase().includes("light") ? payload.lightTextSecondary : payload.textSecondary,
-        borderSoft: payload.themeMode.toLowerCase().includes("light") ? payload.lightBorderSoft : payload.borderSoft,
-      },
-      palettes: {
-        ...(readPath(sourceFranchize, ["theme", "palettes"], {}) as UnknownRecord),
-        dark: {
-          ...(readPath(sourceFranchize, ["theme", "palettes", "dark"], {}) as UnknownRecord),
-          bgBase: payload.bgBase,
-          bgCard: payload.bgCard,
-          accentMain: payload.accentMain,
-          accentMainHover: payload.accentMainHover,
-          textPrimary: payload.textPrimary,
-          textSecondary: payload.textSecondary,
-          borderSoft: payload.borderSoft,
-        },
-        light: {
-          ...(readPath(sourceFranchize, ["theme", "palettes", "light"], {}) as UnknownRecord),
-          bgBase: payload.lightBgBase,
-          bgCard: payload.lightBgCard,
-          accentMain: payload.lightAccentMain,
-          accentMainHover: payload.lightAccentMainHover,
-          textPrimary: payload.lightTextPrimary,
-          textSecondary: payload.lightTextSecondary,
-          borderSoft: payload.lightBorderSoft,
-        },
-      },
-    },
-    header: {
-      ...(readPath(sourceFranchize, ["header"], {}) as UnknownRecord),
-      menuLinks,
-    },
-    footer: {
-      ...(readPath(sourceFranchize, ["footer"], {}) as UnknownRecord),
-      phone: payload.phone,
-      email: payload.email,
-      address: payload.address,
-      socialLinks: parseSocialLinks(payload.socialLinksText),
-    },
-    contacts: {
-      ...(readPath(sourceFranchize, ["contacts"], {}) as UnknownRecord),
-      phone: payload.phone,
-      email: payload.email,
-      address: payload.address,
-      telegram: payload.telegram,
-      map: {
-        ...(readPath(sourceFranchize, ["contacts", "map"], {}) as UnknownRecord),
-        gps: payload.mapGps,
-        imageUrl: payload.mapImageUrl,
-        bounds: {
-          ...(readPath(sourceFranchize, ["contacts", "map", "bounds"], {}) as UnknownRecord),
-          top: Number(payload.mapBoundsTop),
-          bottom: Number(payload.mapBoundsBottom),
-          left: Number(payload.mapBoundsLeft),
-          right: Number(payload.mapBoundsRight),
-        },
-      },
-    },
-    catalog: {
-      ...(readPath(sourceFranchize, ["catalog"], {}) as UnknownRecord),
-      groupOrder: splitCsv(payload.categoryOrderText),
-      promoBanners: parsePromoBanners(payload.promoBannersText, payload.slug),
-      adCards: parseAdCards(payload.adCardsText, payload.slug),
-    },
-    order: {
-      ...(readPath(sourceFranchize, ["order"], {}) as UnknownRecord),
-      allowPromo: payload.allowPromo,
-      deliveryModes: splitCsv(payload.deliveryModesText),
-      paymentOptions: splitCsv(payload.paymentOptionsText),
-      defaultMode: payload.defaultMode,
-    },
-  };
-
-  const mergedMetadata: UnknownRecord = {
-    ...baseMetadata,
-    franchize: franchizeMetadata,
-  };
-
-  const { error: updateError } = await supabaseAdmin.from("crews").update({ metadata: mergedMetadata }).eq("id", crew.id);
-  if (updateError) {
-    logger.error("[franchize] save config failed", updateError);
-    return { ok: false, message: "Не удалось сохранить конфиг в metadata экипажа.", data: payload, canEdit: true };
-  }
-
-  const mergedContractDefaults: UnknownRecord = {
-    ...contractDefaultsOverrides,
-    defaults: {
-      ...(readPath(contractDefaultsOverrides, ["defaults"], {}) as UnknownRecord),
-      issuerName: payload.issuerName || defaultFranchizeConfig.issuerName,
-      issuer_representative: payload.issuerRepresentative || defaultFranchizeConfig.issuerRepresentative,
-      included_mileage: Number(payload.includedMileage) || Number(defaultFranchizeConfig.includedMileage),
-      overage_rate: Number(payload.overageRateRub) || Number(defaultFranchizeConfig.overageRateRub),
-      bike_value_rub: Number(payload.bikeValueRub) || Number(defaultFranchizeConfig.bikeValueRub),
-      bike_value_words: payload.bikeValueWords || defaultFranchizeConfig.bikeValueWords,
-      late_return_penalty_rub: Number(payload.lateReturnPenaltyRub) || Number(defaultFranchizeConfig.lateReturnPenaltyRub),
-      return_address: payload.returnAddress || defaultFranchizeConfig.returnAddress,
-    },
-  };
-
-  await saveCrewSensitiveData(payload.slug, {
-    contractDefaults: mergedContractDefaults,
-    docTemplates: docTemplatesOverrides,
-  });
-
-  return {
-    ok: true,
-    message: `Франшизный конфиг сохранён в crews.metadata.franchize для ${payload.slug}.`,
-    data: payload,
-    canEdit: true,
-  };
-}
-
-function formatMoney(value: number): string {
-  return Number(value || 0).toLocaleString("ru-RU");
-}
-
-const franchizePromoValidationSchema = z.object({
-  slug: z.string().trim().min(1),
-  code: z.string().trim().min(1),
-  baseAmount: z.number().finite().nonnegative(),
-});
-
-type PromoValidationBanner = {
-  title: string;
-  subtitle: string;
-  code: string;
-  activeFrom: string;
-  activeTo: string;
-  ctaLabel: string;
-  discountType: string;
-  discountPercent: number;
-  discountAmount: number;
-  maxDiscount: number;
-  minSubtotal: number;
-};
-
-function normalizePromoCode(value: string): string {
-  return value.trim().replace(/\s+/g, "").toUpperCase();
-}
-
-function parsePromoDate(value: string, boundary: "start" | "end"): Date | null {
-  if (!value.trim()) return null;
-  const date = new Date(boundary === "start" ? `${value}T00:00:00` : `${value}T23:59:59`);
-  return Number.isNaN(date.getTime()) ? null : date;
-}
-
-function isPromoActiveNow(banner: PromoValidationBanner, now = new Date()): boolean {
-  const activeFrom = parsePromoDate(banner.activeFrom, "start");
-  const activeTo = parsePromoDate(banner.activeTo, "end");
-  return (!activeFrom || now >= activeFrom) && (!activeTo || now <= activeTo);
-}
-
-function readPromoNumber(banner: UnknownRecord, keys: string[]): number {
-  for (const key of keys) {
-    const value = readPath<unknown>(banner, [key], undefined);
-    const numericValue = typeof value === "string" ? Number(value.replace(/\s+/g, "")) : Number(value);
-    if (Number.isFinite(numericValue) && numericValue > 0) return numericValue;
-  }
-  return 0;
-}
-
-function capPromoDiscount(amount: number, baseAmount: number, maxDiscount: number): number {
-  const cappedBySubtotal = Math.min(baseAmount, Math.round(amount));
-  return maxDiscount > 0 ? Math.min(cappedBySubtotal, Math.round(maxDiscount)) : cappedBySubtotal;
-}
-
-function extractPromoDiscount(banner: PromoValidationBanner, baseAmount: number): { amount: number; description: string } {
-  if (banner.discountPercent > 0) {
-    const percent = Math.min(90, Math.max(1, banner.discountPercent));
-    const amount = capPromoDiscount((baseAmount * percent) / 100, baseAmount, banner.maxDiscount);
-    return { amount, description: `${percent}% скидка` };
-  }
-
-  if (banner.discountAmount > 0) {
-    const amount = capPromoDiscount(banner.discountAmount, baseAmount, banner.maxDiscount);
-    return { amount, description: `${amount.toLocaleString("ru-RU")} ₽ скидка` };
-  }
-
-  const promoText = [banner.subtitle, banner.title, banner.ctaLabel, banner.code].filter(Boolean).join(" ");
-  const percentMatch = promoText.match(/(?:-|скидка\s*)?(\d{1,2})\s*%/i);
-  if (percentMatch) {
-    const percent = Math.min(90, Math.max(1, Number(percentMatch[1])));
-    const amount = capPromoDiscount((baseAmount * percent) / 100, baseAmount, banner.maxDiscount);
-    return { amount, description: `${percent}% скидка` };
-  }
-
-  const fixedRubMatch = promoText.match(/(?:-|скидка\s*)?(\d[\d\s]{2,})\s*(?:₽|руб\.?)/i);
-  if (fixedRubMatch) {
-    const fixedAmount = Number(fixedRubMatch[1].replace(/\s+/g, ""));
-    if (Number.isFinite(fixedAmount) && fixedAmount > 0) {
-      const amount = capPromoDiscount(fixedAmount, baseAmount, banner.maxDiscount);
-      return { amount, description: `${amount.toLocaleString("ru-RU")} ₽ скидка` };
-    }
-  }
-
-  return { amount: 0, description: banner.subtitle.trim() || banner.title.trim() || "Бонус применён" };
-}
-
-export async function validateFranchizePromoCode(input: unknown): Promise<
-  | { success: true; code: string; title: string; discountAmount: number; description: string }
-  | { success: false; error: string }
-> {
-  const parsed = franchizePromoValidationSchema.safeParse(input);
-  if (!parsed.success) {
-    return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный промокод." };
-  }
-
-  const slug = normalizeCrewSlug(parsed.data.slug);
-  const code = normalizePromoCode(parsed.data.code);
-  const baseAmount = Math.round(parsed.data.baseAmount);
-  if (!slug || !code) return { success: false, error: "Введите промокод из активной акции." };
-  if (baseAmount <= 0) return { success: false, error: "Промокод можно применить после добавления товаров в заказ." };
-
-  const { data: crew, error } = await supabaseAdmin
-    .from("crews")
-    .select("metadata")
-    .eq("slug", slug)
-    .maybeSingle();
-
-  if (error) {
-    logger.error("[franchize] promo validation crew lookup failed", error);
-    return { success: false, error: "Не удалось проверить промокод. Попробуйте ещё раз." };
-  }
-  if (!crew) return { success: false, error: "Витрина для промокода не найдена." };
-
-  const franchize = readPath<UnknownRecord>(crew.metadata, ["franchize"], {});
-  if (!readPath(franchize, ["order", "allowPromo"], true)) {
-    return { success: false, error: "Промокоды отключены для этой витрины." };
-  }
-
-  const availablePromos = readArrayPath<unknown>(franchize, ["catalog", "promoBanners"])
-    .map((item: unknown): PromoValidationBanner => {
-      const banner = (item ?? {}) as UnknownRecord;
-      const discountType = readPath(banner, ["discountType"], readPath(banner, ["discount_type"], "")).toLowerCase();
-      const discountValue = readPromoNumber(banner, ["discountValue", "discount_value"]);
-      return {
-        title: readPath(banner, ["title"], ""),
-        subtitle: readPath(banner, ["subtitle"], ""),
-        code: readPath(banner, ["code"], ""),
-        activeFrom: readPath(banner, ["activeFrom"], ""),
-        activeTo: readPath(banner, ["activeTo"], ""),
-        ctaLabel: readPath(banner, ["ctaLabel"], ""),
-        discountType,
-        discountPercent: readPromoNumber(banner, ["discountPercent", "discount_percent", "percent"]) || (discountType === "percent" ? discountValue : 0),
-        discountAmount: readPromoNumber(banner, ["discountAmount", "discount_amount", "amountRub", "amount_rub"]) || (discountType === "fixed" ? discountValue : 0),
-        maxDiscount: readPromoNumber(banner, ["maxDiscount", "max_discount", "maxDiscountRub", "max_discount_rub"]),
-        minSubtotal: readPromoNumber(banner, ["minSubtotal", "min_subtotal", "minAmount", "min_amount"]),
-      };
-    })
-    .filter((banner) => normalizePromoCode(banner.code).length > 0);
-
-  if (availablePromos.length === 0) return { success: false, error: "Для этой витрины сейчас нет промокодов." };
-
-  const matchedPromo = availablePromos.find((banner) => normalizePromoCode(banner.code) === code);
-  if (!matchedPromo) return { success: false, error: "Промокод не найден. Проверьте код с карточки акции." };
-  if (!isPromoActiveNow(matchedPromo)) return { success: false, error: "Срок действия промокода уже не активен для выбранной витрины." };
-  if (matchedPromo.minSubtotal > 0 && baseAmount < matchedPromo.minSubtotal) {
-    return { success: false, error: `Минимальная сумма для промокода — ${matchedPromo.minSubtotal.toLocaleString("ru-RU")} ₽.` };
-  }
-
-  const discount = extractPromoDiscount(matchedPromo, baseAmount);
-  return {
-    success: true,
-    code,
-    title: matchedPromo.title || code,
-    discountAmount: discount.amount,
-    description: discount.description,
-  };
-}
-
-async function resolveFranchizeCheckoutTotal(
-  payload: z.infer<typeof franchizeOrderInvoiceSchema>,
-): Promise<
-  | { success: true; totalAmount: number; promoCode?: string; promoTitle?: string; promoDiscount: number }
-  | { success: false; error: string }
-> {
-  const baseAmount = Math.round(payload.subtotal + payload.extrasTotal);
-  const code = normalizePromoCode(payload.promoCode ?? "");
-  if (!code) return { success: true, totalAmount: baseAmount, promoDiscount: 0 };
-
-  const result = await validateFranchizePromoCode({ slug: payload.slug, code, baseAmount });
-  if (!result.success) return result;
-
-  const expectedDiscount = Math.min(result.discountAmount, baseAmount);
-  const submittedDiscount = Math.min(Math.round(payload.promoDiscount ?? 0), baseAmount);
-  if (submittedDiscount !== expectedDiscount) {
-    return { success: false, error: "Сумма промокода изменилась. Примените промокод ещё раз перед отправкой." };
-  }
-
-  return {
-    success: true,
-    totalAmount: Math.max(0, baseAmount - expectedDiscount),
-    promoCode: result.code,
-    promoTitle: result.title,
-    promoDiscount: expectedDiscount,
-  };
-}
-
-function parseDurationDays(rawDuration: string): number {
-  const normalized = String(rawDuration || "").toLowerCase();
-  const numericMatch = normalized.match(/\d+/);
-  const days = numericMatch ? Number(numericMatch[0]) : 1;
-  if (!Number.isFinite(days) || days <= 0) {
-    return 1;
-  }
-  return days;
-}
-
-type FranchizeOrderNotifyPayload = z.infer<typeof franchizeOrderInvoiceSchema> & {
-  totalAmount: number;
-  subtotal: number;
-  extrasTotal: number;
-};
-
-type FranchizeOrderFlowType = "rental" | "sale" | "mixed";
-
-async function loadFranchizeDealTemplate(slug: string, flowType: FranchizeOrderFlowType): Promise<string> {
-  const crewSensitive = await getCrewSensitiveData(slug);
-  const secureTemplateKey = flowType === "rental" ? "rentalDealTemplate" : "saleDealTemplate";
-  const secureTemplate = readPath(crewSensitive.docTemplates ?? {}, [secureTemplateKey], "");
-  if (typeof secureTemplate === "string" && secureTemplate.trim().length > 0) {
-    return secureTemplate;
-  }
-
-  const remoteTemplateUrl = flowType === "rental"
-    ? "https://raw.githubusercontent.com/salavey13/carTest/main/docs/RENTAL_DEAL_TEMPLATE_DEMO.md"
-    : "https://raw.githubusercontent.com/salavey13/carTest/main/docs/SALE_DEAL_TEMPLATE_DEMO.md";
-  const localTemplateFile = flowType === "rental" ? "RENTAL_DEAL_TEMPLATE_DEMO.md" : "SALE_DEAL_TEMPLATE_DEMO.md";
-
-  try {
-    const response = await fetch(remoteTemplateUrl, { cache: "no-store" });
-    if (!response.ok) {
-      throw new Error(`template fetch failed with ${response.status}`);
-    }
-
-    const remoteTemplate = await response.text();
-    if (remoteTemplate.trim().length > 0) {
-      return remoteTemplate;
-    }
-
-    throw new Error("template fetch returned empty content");
-  } catch (error) {
-    logger.warn("[franchize] fallback to local template", { flowType, error });
-    const localTemplatePath = path.join(process.cwd(), "docs", localTemplateFile);
-    return readFile(localTemplatePath, "utf8");
-  }
-}
-
-async function createFranchizeOrderNotificationLog(payload: FranchizeOrderNotifyPayload) {
-  const snapshot = {
-    ...payload,
-    persistedAt: new Date().toISOString(),
-  };
-
-  const { data, error } = await supabaseAdmin
-    .from("franchize_order_notifications")
-    .insert({
-      slug: payload.slug,
-      order_id: payload.orderId,
-      payload: snapshot,
-      send_status: "pending",
-      attempts: 1,
-    })
-    .select("id")
-    .single();
-
-  if (error) {
-    throw new Error(`Failed to persist order snapshot: ${error.message}`);
-  }
-
-  return data.id as string;
-}
-
-async function updateFranchizeOrderNotificationLog(
-  logId: string,
-  patch: {
-    send_status: "sent" | "failed";
-    rendered_markdown?: string;
-    doc_file_name?: string;
-    last_error?: string;
-  },
-) {
-  const { error } = await supabaseAdmin
-    .from("franchize_order_notifications")
-    .update({
-      ...patch,
-      updated_at: new Date().toISOString(),
-    })
-    .eq("id", logId);
-
-  if (error) {
-    logger.error("[franchize] failed to update franchize_order_notifications", { logId, error });
-  }
-}
-
-async function buildFranchizeOrderDocAndNotify(payload: FranchizeOrderNotifyPayload) {
-  const logId = await createFranchizeOrderNotificationLog(payload);
-
-  try {
-    const { data: cars, error: carsError } = await supabaseAdmin
-      .from("cars")
-      .select("id, make, model, specs")
-      .in("id", payload.cartLines.map((line) => line.itemId));
-
-    if (carsError) {
-      logger.error("[franchize] failed to load cars for order doc", carsError);
-    }
-
-    const byId = new Map((cars || []).map((car) => [car.id, car]));
-    const extrasRows = payload.extras.length
-      ? payload.extras
-          .map((extra) => `| ${extra.label} | 1 | ${formatMoney(extra.amount)} ₽ | ${formatMoney(extra.amount)} ₽ |`)
-          .join("\n")
-      : "| Без доп. опций | 0 | 0 ₽ | 0 ₽ |";
-
-    const firstLine = payload.cartLines[0];
-    const firstCar = firstLine ? byId.get(firstLine.itemId) : null;
-    const firstSpecs = (firstCar?.specs as Record<string, unknown> | null) || {};
-    const rentDays = Math.max(...payload.cartLines.map((line) => parseDurationDays(line.options.duration)));
-    const flowType: FranchizeOrderFlowType = payload.flowType ?? "rental";
-    const isSaleFlow = flowType === "sale" || flowType === "mixed";
-    const rentStartDate = payload.rentalStartDate || payload.time;
-    const rentEndDate = payload.rentalEndDate || payload.time;
-    const dailyPriceRub = firstLine?.pricePerDay || Math.round(payload.subtotal / Math.max(1, rentDays));
-
-    const template = await loadFranchizeDealTemplate(payload.slug, flowType);
-    const userSensitive = await getUserSensitiveData(payload.telegramUserId);
-    const crewSensitive = await getCrewSensitiveData(payload.slug);
-    const contractDefaults = (crewSensitive.contractDefaults ?? {}) as UnknownRecord;
-    const defaults = (readPath(contractDefaults, ["defaults"], {}) ?? {}) as UnknownRecord;
-
-    const variables = {
-      contract_number: `${payload.slug.toUpperCase()}-${payload.orderId}`,
-      contract_date: new Date().toLocaleDateString("ru-RU"),
-      day: new Date().getDate().toString().padStart(2, "0"),
-      month: new Date().toLocaleString("ru-RU", { month: "long" }),
-      year: new Date().getFullYear().toString(),
-      renter_full_name: payload.recipient,
-      renter_phone: payload.phone,
-      renter_driver_license: userSensitive.driverLicense || payload.renterDriverLicense || "указывается при выдаче",
-      renter_passport: userSensitive.passport || payload.renterPassport || "указывается при выдаче",
-      issuer_name: String(readPath(defaults, ["issuerName"], `Franchize ${payload.slug}`)),
-      issuer_signatory: "Администратор экипажа",
-      issuer_representative: String(readPath(defaults, ["issuer_representative"], "Сидоров Илья Олегович")),
-      bike_make_model: firstCar
-        ? `${firstCar.make || "Bike"} ${firstCar.model || "Model"}`
-        : payload.cartLines.map((line) => line.itemId).join(", "),
-      bike_vin: String(firstSpecs.vin || firstSpecs.frame || firstSpecs.vin_number || "уточняется"),
-      bike_plate: String(firstSpecs.plate || firstSpecs.state_number || "уточняется"),
-      bike_mileage: "45073",
-      rent_start_time: "12:00",
-      rent_start_date: rentStartDate,
-      rent_end_time: "12:00",
-      rent_end_date: rentEndDate,
-      rent_days: rentDays,
-      daily_price_rub: formatMoney(dailyPriceRub),
-      subtotal_rub: formatMoney(payload.subtotal),
-      extras_rows: extrasRows,
-      extras_total_rub: formatMoney(payload.extrasTotal),
-      total_price_rub: formatMoney(payload.totalAmount),
-      deposit_rub: "20 000",
-      included_mileage: String(readPath(defaults, ["included_mileage"], 200)),
-      overage_rate: `${readPath(defaults, ["overage_rate"], 30)} руб/км`,
-      bike_value_rub: String(readPath(defaults, ["bike_value_rub"], 700000)),
-      bike_value_words: String(readPath(defaults, ["bike_value_words"], "Семьсот тысяч")),
-      late_return_penalty_rub: String(readPath(defaults, ["late_return_penalty_rub"], 5000)),
-      return_address: String(readPath(defaults, ["return_address"], "г. Нижний Новгород, ул. Стригинский переулок, дом 13б")),
-      signature_timestamp: new Date().toLocaleString("ru-RU"),
-      signature_fingerprint: payload.signatureFingerprint || "—",
-      renter_signature: payload.signatureName || "электронное согласие в Telegram WebApp",
-      document_key: `${isSaleFlow ? "sale" : "rental"}-${payload.slug}-${payload.orderId}`,
-      verified_at: new Date().toISOString(),
-    } as const;
-
-    const docFileName = `franchize-order-${payload.slug}-${payload.orderId}.docx`;
-    const verifierScope = `${flowType}:${payload.slug}:${payload.orderId}`;
-    const { bytes, renderedMarkdown, verifierRecordId, sha256 } = await buildFranchizeDocxFromTemplate({
-      integrationScope: verifierScope,
-      uploadedBy: "franchize-order-system",
-      documentKey: variables.document_key,
-      fileName: docFileName,
-      template,
-      variables,
-    });
-
-    const adminChatId = process.env.ADMIN_CHAT_ID;
-    if (!adminChatId) {
-      throw new Error("ADMIN_CHAT_ID not configured");
-    }
-
-    await notifyAdmin(
-      [
-        `${isSaleFlow ? "🛍️ Новый franchize-заказ на покупку" : "🧾 Новый franchize-заказ на аренду"} #${payload.orderId}`,
-        `Crew: ${payload.slug}`,
-        `Получатель: ${payload.recipient}`,
-        `Телефон: ${payload.phone}`,
-        `Время: ${payload.time}`,
-        `Оплата: ${payload.payment}`,
-        `Доставка: ${payload.delivery}`,
-        `Итого: ${formatMoney(payload.totalAmount)} ₽`,
-      ].join("\n"),
-    );
-
-    const recipientSet = new Set<string>([adminChatId, payload.telegramUserId]);
-    const { data: crewRow } = await supabaseAdmin.from("crews").select("owner_id").eq("slug", payload.slug).maybeSingle();
-    const ownerId = typeof crewRow?.owner_id === "string" ? crewRow.owner_id : "";
-    if (ownerId) {
-      const { data: ownerUser } = await supabaseAdmin
-        .from("users")
-        .select("metadata")
-        .eq("user_id", ownerId)
-        .maybeSingle();
-      const ownerMeta = (ownerUser?.metadata ?? {}) as Record<string, unknown>;
-      const ownerTelegramId = String(ownerMeta.telegram_id ?? ownerMeta.telegramId ?? "").trim();
-      if (ownerTelegramId) {
-        recipientSet.add(ownerTelegramId);
-      }
-    }
-
-    for (const recipientId of recipientSet) {
-      const sendDocResult = await sendTelegramDocument(recipientId, new Blob([bytes]), docFileName);
-      if (!sendDocResult.success) {
-        throw new Error(sendDocResult.error || `Failed to send DOCX to ${recipientId}`);
-      }
-    }
-
-    await updateFranchizeOrderNotificationLog(logId, {
-      send_status: "sent",
-      rendered_markdown: renderedMarkdown,
-      doc_file_name: docFileName,
-      last_error: "",
-    });
-
-    if (flowType === "rental") {
-      const { data: rentalRow, error: rentalLookupError } = await supabaseAdmin
-        .from("rentals")
-        .select("rental_id, metadata")
-        .eq("metadata->>orderId", payload.orderId)
-        .order("created_at", { ascending: false })
-        .limit(1)
-        .maybeSingle();
-
-      if (rentalLookupError) {
-        logger.warn("[franchize] rental lookup for contract attachment failed", {
-          orderId: payload.orderId,
-          slug: payload.slug,
-          error: rentalLookupError.message,
-        });
-      } else if (!rentalRow?.rental_id) {
-        logger.warn("[franchize] rental not found for contract attachment", {
-          orderId: payload.orderId,
-          slug: payload.slug,
-          verifierScope,
-          documentKey: variables.document_key,
-        });
-      } else {
-        const existingMetadata =
-          rentalRow.metadata && typeof rentalRow.metadata === "object" ? (rentalRow.metadata as Record<string, unknown>) : {};
-        const rentalScope = `rental:${rentalRow.rental_id}`;
-        const { error: rentalUpdateError } = await supabaseAdmin
-          .from("rentals")
-          .update({
-            metadata: {
-              ...existingMetadata,
-              contract_verifier: {
-                scope: rentalScope,
-                sourceScope: verifierScope,
-                documentKey: variables.document_key,
-                docVerifierRecordId: verifierRecordId ?? null,
-                originalSha256: sha256,
-                status: "verified",
-                verifiedAt: new Date().toISOString(),
-                expiresAt: null,
-              },
-            },
-          })
-          .eq("rental_id", rentalRow.rental_id);
-
-        if (rentalUpdateError) {
-          logger.warn("[franchize] rental contract verifier attachment failed", {
-            rentalId: rentalRow.rental_id,
-            orderId: payload.orderId,
-            error: rentalUpdateError.message,
-          });
-        }
-      }
-    }
-
-    return { renderedMarkdown, docFileName };
-  } catch (error) {
-    const message = error instanceof Error ? error.message : "Failed to submit order notification";
-    await updateFranchizeOrderNotificationLog(logId, {
-      send_status: "failed",
-      last_error: message,
-    });
-
-    throw error;
-  }
-}
-
-export async function reconcileRentalContractVerifierAttachment(input: {
-  rentalId: string;
-  orderId?: string;
-  flowType?: "rental" | "sale" | "mixed";
-  slug?: string;
-}): Promise<{ success: boolean; error?: string; attached?: boolean; recordId?: string }> {
-  const rentalId = input.rentalId.trim();
-  if (!rentalId) return { success: false, error: "rentalId is required" };
-
-  const { data: rental, error: rentalError } = await supabaseAdmin
-    .from("rentals")
-    .select("rental_id, metadata")
-    .eq("rental_id", rentalId)
-    .maybeSingle();
-
-  if (rentalError || !rental) {
-    return { success: false, error: rentalError?.message || "rental not found" };
-  }
-
-  const metadata = rental.metadata && typeof rental.metadata === "object" ? (rental.metadata as Record<string, unknown>) : {};
-  const verifier = metadata.contract_verifier && typeof metadata.contract_verifier === "object"
-    ? (metadata.contract_verifier as Record<string, unknown>)
-    : null;
-
-  const flowType = input.flowType ?? "rental";
-  const slug = (input.slug || String(metadata.slug || "")).trim();
-  const orderId = (input.orderId || String(metadata.orderId || "")).trim();
-
-  const sourceScope = typeof verifier?.sourceScope === "string" && verifier.sourceScope.trim()
-    ? verifier.sourceScope.trim()
-    : slug && orderId
-      ? `${flowType}:${slug}:${orderId}`
-      : "";
-  const keyPrefix = flowType === "sale" || flowType === "mixed" ? "sale" : "rental";
-  const documentKey = typeof verifier?.documentKey === "string" && verifier.documentKey.trim()
-    ? verifier.documentKey.trim()
-    : slug && orderId
-      ? `${keyPrefix}-${slug}-${orderId}`
-      : "";
-
-  if (!sourceScope || !documentKey) {
-    return { success: false, error: "Cannot derive sourceScope/documentKey for reconciliation" };
-  }
-
-  const { data: record, error: recordError } = await supabaseAdmin
-    .from("doc_verifier_records")
-    .select("id, integration_scope, document_key, original_sha256")
-    .eq("integration_scope", sourceScope)
-    .eq("document_key", documentKey)
-    .maybeSingle();
-
-  if (recordError || !record) {
-    return { success: false, error: recordError?.message || "doc_verifier record not found" };
-  }
-
-  const contractVerifierPatch = {
-    ...(verifier ?? {}),
-    scope: `rental:${rentalId}`,
-    sourceScope: sourceScope,
-    documentKey: documentKey,
-    docVerifierRecordId: record.id,
-    originalSha256: record.original_sha256,
-    status: "verified",
-    verifiedAt: new Date().toISOString(),
-    expiresAt: typeof verifier?.expiresAt === "string" ? verifier.expiresAt : null,
-  };
-
-  const { error: updateError } = await supabaseAdmin
-    .from("rentals")
-    .update({
-      metadata: {
-        ...metadata,
-        contract_verifier: contractVerifierPatch,
-      },
-    })
-    .eq("rental_id", rentalId);
-
-  if (updateError) return { success: false, error: updateError.message };
-  return { success: true, attached: true, recordId: record.id };
-}
-
-
-const franchizeOrderInvoiceSchema = z.object({
-  slug: z.string().trim().min(1),
-  orderId: z.string().trim().min(1),
-  telegramUserId: z.string().trim().min(1),
-  recipient: z.string().trim().min(2),
-  phone: z.string().trim().min(6),
-  renterDriverLicense: z.string().trim().optional(),
-  renterPassport: z.string().trim().optional(),
-  time: z.string().trim().min(1),
-  comment: z.string().trim().default(""),
-  rentalStartDate: z.string().trim().optional(),
-  rentalEndDate: z.string().trim().optional(),
-  signatureName: z.string().trim().optional(),
-  signatureAccepted: z.boolean().optional(),
-  signatureFingerprint: z.string().trim().optional(),
-  payment: z.enum(["telegram_xtr", "card", "cash", "sbp"]),
-  delivery: z.enum(["pickup", "delivery"]),
-  subtotal: z.number().finite().nonnegative(),
-  extrasTotal: z.number().finite().nonnegative().default(0),
-  promoCode: z.string().trim().optional(),
-  promoTitle: z.string().trim().optional(),
-  promoDiscount: z.number().finite().nonnegative().default(0),
-  totalAmount: z.number().finite().nonnegative().default(0),
-  extras: z
-    .array(
-      z.object({
-        id: z.string().trim().min(1),
-        label: z.string().trim().min(1),
-        amount: z.number().finite().nonnegative(),
-      }),
-    )
-    .default([]),
-  cartLines: z.array(z.object({
-    lineId: z.string().trim().min(1).optional(),
-    itemId: z.string().trim().min(1),
-    qty: z.number().int().positive(),
-    pricePerDay: z.number().finite().nonnegative(),
-    lineTotal: z.number().finite().nonnegative(),
-    options: z
-      .object({
-        package: z.string().trim().default("Базовый"),
-        duration: z.string().trim().default("1 день"),
-        perk: z.string().trim().default("Стандарт"),
-        auction: z.string().trim().default("Без аукциона"),
-      })
-      .default({ package: "Базовый", duration: "1 день", perk: "Стандарт", auction: "Без аукциона" }),
-  })).min(1),
-  flowType: z.enum(["rental", "sale", "mixed"]).default("rental"),
-});
-
-export async function submitFranchizeOrderNotification(input: unknown): Promise<{ success: boolean; error?: string }> {
-  const parsed = franchizeOrderInvoiceSchema.safeParse(input);
-
-  if (!parsed.success) {
-    return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный payload заказа." };
-  }
-
-  const payload = parsed.data;
-  const totalResult = await resolveFranchizeCheckoutTotal(payload);
-  if (!totalResult.success) return totalResult;
-  const effectiveTotal = totalResult.totalAmount;
-  const validatedPayload = { ...payload, ...totalResult, totalAmount: effectiveTotal };
-
-  try {
-    await buildFranchizeOrderDocAndNotify({ ...validatedPayload, totalAmount: effectiveTotal, subtotal: payload.subtotal, extrasTotal: payload.extrasTotal });
-    return { success: true };
-  } catch (error) {
-    logger.error("[franchize] submitFranchizeOrderNotification failed", error);
-    return { success: false, error: error instanceof Error ? error.message : "Failed to submit order notification" };
-  }
-}
-
-const retryFranchizeNotificationSchema = z.object({
-  slug: z.string().trim().min(1),
-  orderId: z.string().trim().min(1),
-  actorUserId: z.string().trim().min(1),
-});
-
-export async function retryFranchizeOrderNotification(input: unknown): Promise<{ success: boolean; error?: string }> {
-  const parsed = retryFranchizeNotificationSchema.safeParse(input);
-  if (!parsed.success) {
-    return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный payload для retry." };
-  }
-
-  const { slug, orderId, actorUserId } = parsed.data;
-  const { data: crew } = await supabaseAdmin.from("crews").select("id, owner_id").eq("slug", slug).maybeSingle();
-  if (!crew) return { success: false, error: "Экипаж не найден." };
-  const canRetry = await resolveFranchizeEditorAccess(actorUserId, crew);
-  if (!canRetry) return { success: false, error: "Недостаточно прав для retry уведомления." };
-  const { data: logRow, error } = await supabaseAdmin
-    .from("franchize_order_notifications")
-    .select("payload")
-    .eq("slug", slug)
-    .eq("order_id", orderId)
-    .order("created_at", { ascending: false })
-    .limit(1)
-    .maybeSingle();
-
-  if (error) {
-    return { success: false, error: `Не удалось загрузить snapshot для retry: ${error.message}` };
-  }
-
-  if (!logRow?.payload || typeof logRow.payload !== "object") {
-    return { success: false, error: "Snapshot заказа не найден для retry." };
-  }
-
-  return submitFranchizeOrderNotification(logRow.payload);
-}
-
-export async function getFranchizeOrderNotificationFailures(input: unknown): Promise<{ success: boolean; items?: Array<{ orderId: string; sendTo: string; lastError: string; createdAt: string }>; error?: string }> {
-  const parsed = z.object({ slug: z.string().trim().min(1), actorUserId: z.string().trim().min(1) }).safeParse(input);
-  if (!parsed.success) return { success: false, error: "Некорректный запрос." };
-  const { slug, actorUserId } = parsed.data;
-  const { data: crew } = await supabaseAdmin.from("crews").select("id, owner_id").eq("slug", slug).maybeSingle();
-  if (!crew) return { success: false, error: "Экипаж не найден." };
-  const canRead = await resolveFranchizeEditorAccess(actorUserId, crew);
-  if (!canRead) return { success: false, error: "Недостаточно прав для просмотра логов." };
-
-  const { data, error } = await supabaseAdmin
-    .from("franchize_order_notifications")
-    .select("order_id, send_to, last_error, created_at")
-    .eq("slug", slug)
-    .eq("send_status", "failed")
-    .order("created_at", { ascending: false })
-    .limit(20);
-  if (error) return { success: false, error: error.message };
-  return {
-    success: true,
-    items: (data ?? []).map((row) => ({ orderId: row.order_id, sendTo: row.send_to ?? "", lastError: row.last_error ?? "", createdAt: row.created_at ?? "" })),
-  };
-}
-
-const rentalReviewInputSchema = z.object({
-  slug: z.string().trim().min(1),
-  rentalId: z.string().uuid(),
-  initData: z.string().trim().min(1),
-  rating: z.coerce.number().int().min(1).max(5),
-  text: z.string().trim().max(1200).default(""),
-});
-
-async function resolveVerifiedTelegramUserId(initDataString: string): Promise<{ success: boolean; userId?: string; error?: string }> {
-  const botToken = process.env.TELEGRAM_BOT_TOKEN;
-  if (!botToken) return { success: false, error: "Telegram bot token is not configured." };
-
-  const params = new URLSearchParams(initDataString);
-  const hashFromClient = params.get("hash");
-  if (!hashFromClient) return { success: false, error: "Telegram init data hash is missing." };
-
-  const keys = Array.from(params.keys()).filter((key) => key !== "hash" && key !== "signature").sort();
-  const dataCheckString = keys.map((key) => `${key}=${params.get(key)}`).join("\n");
-
-  const botTokenKey = await webcrypto.subtle.importKey(
-    "raw",
-    new TextEncoder().encode(botToken),
-    { name: "HMAC", hash: "SHA-256" },
-    false,
-    ["sign"],
-  );
-  const derivedKeyBuffer = await webcrypto.subtle.sign("HMAC", botTokenKey, new TextEncoder().encode("WebAppData"));
-  const telegramWebAppKey = await webcrypto.subtle.importKey(
-    "raw",
-    derivedKeyBuffer,
-    { name: "HMAC", hash: "SHA-256" },
-    false,
-    ["sign"],
-  );
-  const signatureBuffer = await webcrypto.subtle.sign("HMAC", telegramWebAppKey, new TextEncoder().encode(dataCheckString));
-  const signatureHex = Array.from(new Uint8Array(signatureBuffer)).map((byte) => byte.toString(16).padStart(2, "0")).join("");
-  const bypassValidation = process.env.TEMP_BYPASS_TG_AUTH_VALIDATION === "true";
-
-  if (signatureHex !== hashFromClient && !bypassValidation) {
-    return { success: false, error: "Telegram init data hash mismatch." };
-  }
-
-  const userParam = params.get("user");
-  if (!userParam) return { success: false, error: "Telegram init data user is missing." };
-
-  const authDate = Number(params.get("auth_date") ?? 0);
-  const maxInitDataAgeSeconds = 24 * 60 * 60;
-  if (!bypassValidation && (!Number.isFinite(authDate) || authDate <= 0 || Date.now() / 1000 - authDate > maxInitDataAgeSeconds)) {
-    return { success: false, error: "Telegram init data is expired." };
-  }
-
-  try {
-    const user = JSON.parse(userParam) as { id?: unknown };
-    const userId = typeof user.id === "number" || typeof user.id === "string" ? String(user.id) : "";
-    if (!userId) return { success: false, error: "Telegram init data user id is missing." };
-    return { success: true, userId };
-  } catch {
-    return { success: false, error: "Telegram init data user payload is invalid." };
-  }
-}
-
-export async function getRentalReviewContext(input: unknown): Promise<{ success: boolean; error?: string; rental?: { rentalId: string; status: string; userId: string; bikeId: string; bikeTitle: string; crewId: string; existingReview?: RentalReviewVM } }> {
-  const parsed = z.object({ slug: z.string().trim().min(1), rentalId: z.string().uuid() }).safeParse(input);
-  if (!parsed.success) return { success: false, error: "Некорректная ссылка отзыва." };
-  const { slug, rentalId } = parsed.data;
-
-  const { data: crew } = await supabaseAdmin.from("crews").select("id").eq("slug", normalizeCrewSlug(slug)).maybeSingle();
-  if (!crew) return { success: false, error: "Экипаж не найден." };
-
-  const { data: rental, error } = await supabaseAdmin
-    .from("rentals")
-    .select("rental_id, user_id, vehicle_id, status, vehicle:cars(id, make, model, crew_id)")
-    .eq("rental_id", rentalId)
-    .maybeSingle();
-  if (error) return { success: false, error: error.message };
-  if (!rental) return { success: false, error: "Аренда не найдена." };
-
-  const vehicle = Array.isArray(rental.vehicle) ? rental.vehicle[0] : rental.vehicle;
-  if (vehicle?.crew_id !== crew.id) return { success: false, error: "Аренда относится к другому экипажу." };
-
-  const { data: existing } = await supabaseAdmin
-    .from("rental_reviews")
-    .select("id, rental_id, user_id, bike_id, crew_id, rating, text, created_at, hidden_at")
-    .eq("rental_id", rentalId)
-    .maybeSingle();
-
-  return {
-    success: true,
-    rental: {
-      rentalId: String(rental.rental_id),
-      status: String(rental.status ?? ""),
-      userId: String(rental.user_id ?? ""),
-      bikeId: String(rental.vehicle_id ?? ""),
-      bikeTitle: `${vehicle?.make ?? ""} ${vehicle?.model ?? ""}`.trim() || "байк",
-      crewId: String(crew.id),
-      existingReview: existing && !(existing as UnknownRecord).hidden_at ? toRentalReviewVM(existing as UnknownRecord) : undefined,
-    },
-  };
-}
-
-export async function submitRentalReview(input: unknown): Promise<{ success: boolean; error?: string }> {
-  const parsed = rentalReviewInputSchema.safeParse(input);
-  if (!parsed.success) return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный отзыв." };
-  const { slug, rentalId, initData, rating, text } = parsed.data;
-
-  const verifiedTelegramUser = await resolveVerifiedTelegramUserId(initData);
-  if (!verifiedTelegramUser.success || !verifiedTelegramUser.userId) {
-    return { success: false, error: verifiedTelegramUser.error ?? "Не удалось подтвердить Telegram-пользователя." };
-  }
-
-  const context = await getRentalReviewContext({ slug, rentalId });
-  if (!context.success || !context.rental) return { success: false, error: context.error ?? "Аренда не найдена." };
-  if (context.rental.status !== "completed") return { success: false, error: "Отзыв можно оставить после завершения аренды." };
-  if (context.rental.userId !== verifiedTelegramUser.userId) return { success: false, error: "Отзыв может оставить только арендатор." };
-
-  const { data: existingReview, error: existingReviewError } = await supabaseAdmin
-    .from("rental_reviews")
-    .select("hidden_at")
-    .eq("rental_id", rentalId)
-    .maybeSingle();
-  if (existingReviewError) return { success: false, error: existingReviewError.message };
-  if (existingReview?.hidden_at) {
-    return { success: false, error: "Отзыв скрыт модератором. Обновление возможно только после восстановления владельцем экипажа." };
-  }
-
-  const { error } = await supabaseAdmin.from("rental_reviews").upsert(
-    {
-      rental_id: rentalId,
-      user_id: verifiedTelegramUser.userId,
-      bike_id: context.rental.bikeId,
-      crew_id: context.rental.crewId,
-      rating,
-      text,
-    },
-    { onConflict: "rental_id" },
-  );
-
-  if (error) return { success: false, error: error.message };
-  return { success: true };
-}
-
-export async function getFranchizeRentalReviewsForModeration(input: unknown): Promise<{ success: boolean; error?: string; items?: RentalReviewVM[] }> {
-  const parsed = z.object({ slug: z.string().trim().min(1), actorUserId: z.string().trim().min(1) }).safeParse(input);
-  if (!parsed.success) return { success: false, error: "Некорректный запрос." };
-  const { slug, actorUserId } = parsed.data;
-  const { data: crew } = await supabaseAdmin.from("crews").select("id, owner_id").eq("slug", normalizeCrewSlug(slug)).maybeSingle();
-  if (!crew) return { success: false, error: "Экипаж не найден." };
-  const canRead = await resolveFranchizeEditorAccess(actorUserId, crew);
-  if (!canRead) return { success: false, error: "Недостаточно прав для модерации отзывов." };
-
-  const { data, error } = await supabaseAdmin
-    .from("rental_reviews")
-    .select("id, rental_id, user_id, bike_id, crew_id, rating, text, created_at, hidden_at")
-    .eq("crew_id", crew.id)
-    .order("created_at", { ascending: false })
-    .limit(40);
-  if (error) return { success: false, error: error.message };
-  return { success: true, items: ((data ?? []) as UnknownRecord[]).map(toRentalReviewVM) };
-}
-
-export async function moderateRentalReview(input: unknown): Promise<{ success: boolean; error?: string }> {
-  const parsed = z.object({ slug: z.string().trim().min(1), actorUserId: z.string().trim().min(1), reviewId: z.string().uuid(), hidden: z.boolean() }).safeParse(input);
-  if (!parsed.success) return { success: false, error: "Некорректный запрос модерации." };
-  const { slug, actorUserId, reviewId, hidden } = parsed.data;
-  const { data: crew } = await supabaseAdmin.from("crews").select("id, owner_id").eq("slug", normalizeCrewSlug(slug)).maybeSingle();
-  if (!crew) return { success: false, error: "Экипаж не найден." };
-  const canModerate = await resolveFranchizeEditorAccess(actorUserId, crew);
-  if (!canModerate) return { success: false, error: "Недостаточно прав для модерации отзывов." };
-
-  const { error } = await supabaseAdmin
-    .from("rental_reviews")
-    .update({ hidden_at: hidden ? new Date().toISOString() : null, moderated_by: actorUserId })
-    .eq("id", reviewId)
-    .eq("crew_id", crew.id);
-  if (error) return { success: false, error: error.message };
-  return { success: true };
-}
-
-type FranchizeOrderInvoicePayload = z.infer<typeof franchizeOrderInvoiceSchema>;
-
-const franchizeCheckoutRateLimits = new Map<string, number>();
-const FRANCHIZE_CHECKOUT_COOLDOWN_MS = 8_000;
-
-function buildFranchizeInvoiceId(payload: FranchizeOrderInvoicePayload) {
-  const safeSlug = payload.slug.replace(/[^a-z0-9-]/gi, "-").toLowerCase();
-  const safeOrderId = payload.orderId.replace(/[^a-z0-9-]/gi, "-").toLowerCase();
-  const safeUserId = payload.telegramUserId.replace(/[^a-z0-9-]/gi, "-").toLowerCase();
-  return `franchize_order_${safeSlug}_${safeOrderId}_${safeUserId}`;
-}
-
-async function cleanupPendingFranchizeInvoiceAfterSendFailure(invoiceId: string, rentalId: string) {
-  const { error } = await supabaseAdmin
-    .from("invoices")
-    .delete()
-    .eq("id", invoiceId)
-    .eq("status", "pending")
-    .contains("metadata", { rental_id: rentalId });
-
-  if (error) {
-    logger.error("[franchize] failed to cleanup unsent invoice", {
-      invoiceId,
-      rentalId,
-      error: error.message,
-    });
-    return false;
-  }
-
-  logger.warn("[franchize] cleaned up unsent pending invoice after Telegram XTR send failure", {
-    invoiceId,
-    rentalId,
-  });
-  return true;
-}
-
-async function createFranchizeOrderInvoiceInternal(
-  payload: FranchizeOrderInvoicePayload,
-  options: { skipNotification?: boolean } = {},
-): Promise<{ success: boolean; invoiceId?: string; amountXtr?: number; error?: string }> {
-  if (payload.payment !== "telegram_xtr") {
-    return {
-      success: false,
-      error: "Для этой операции поддерживается только Telegram Stars (XTR).",
-    };
-  }
-
-  const effectiveTotal = payload.totalAmount > 0 ? payload.totalAmount : payload.subtotal + payload.extrasTotal;
-  const flowType = payload.flowType ?? "rental";
-  let amountXtr: number;
-  let invoiceTitle = "Franchize: подтверждение намерения";
-  let invoiceDescriptionPrefix = "Подтверждение аренды";
-  let invoiceProofLabel = "Proof-of-interest tip";
-
-  if (flowType === "sale") {
-    amountXtr = Math.min(500, Math.max(100, Math.ceil(effectiveTotal * 0.001)));
-    invoiceTitle = "Franchize: Бронь тест-драйва";
-    invoiceDescriptionPrefix = "Резерв на тест-драйв и получение спеццены на";
-    invoiceProofLabel = "Бронь тест-драйва";
-  } else {
-    amountXtr = Math.max(1, Math.ceil(effectiveTotal * 0.01));
-  }
-
-  const invoiceId = buildFranchizeInvoiceId(payload);
-
-  const { data: existingInvoice } = await supabaseAdmin
-    .from("invoices")
-    .select("id, amount, status")
-    .eq("id", invoiceId)
-    .maybeSingle();
-  if (existingInvoice?.id) {
-    return { success: true, invoiceId, amountXtr: Number(existingInvoice.amount) || amountXtr };
-  }
-
-  const rentalId = randomUUID();
-  const startParamPrefix = flowType === "rental" ? "rental" : "sale";
-  const startParam = `${startParamPrefix}-${rentalId}`;
-  const telegramWebappLink = `https://t.me/oneBikePlsBot/app?startapp=${startParam}`;
-  const siteBaseUrl = process.env.NEXT_PUBLIC_SITE_URL || "https://v0-car-test.vercel.app";
-  const franchizeRentalLink = `${siteBaseUrl}/franchize/${payload.slug}/rental/${rentalId}`;
-
-  const cartText = payload.cartLines
-    .map((line) => `• ${line.itemId} × ${line.qty} (${line.options.package}, ${line.options.duration}, ${line.options.perk}, ${line.options.auction})`)
-    .join("\n");
-
-  const description = [
-    `${invoiceDescriptionPrefix}: ${payload.slug}`,
-    `Заказ: #${payload.orderId}`,
-    `Получатель: ${payload.recipient}`,
-    `Телефон: ${payload.phone}`,
-    `Время: ${payload.time}`,
-    `Получение: ${payload.delivery === "pickup" ? "Самовывоз" : "Доставка"}`,
-    `Состав:`,
-    cartText,
-    `Subtotal: ${payload.subtotal.toLocaleString("ru-RU")} ₽`,
-    `Extras: ${payload.extrasTotal.toLocaleString("ru-RU")} ₽`,
-    payload.promoCode ? `Promo: ${payload.promoCode} (-${payload.promoDiscount.toLocaleString("ru-RU")} ₽)` : "",
-    `Total: ${effectiveTotal.toLocaleString("ru-RU")} ₽`,
-    `${invoiceProofLabel}: ${amountXtr} XTR${flowType === "sale" ? "" : " (1%)"}`,
-    `WebApp: ${telegramWebappLink}`,
-    `Franchize card: ${franchizeRentalLink}`,
-  ]
-    .filter(Boolean)
-    .join("\n");
-
-  const metadata = {
-    slug: payload.slug,
-    orderId: payload.orderId,
-    recipient: payload.recipient,
-    phone: payload.phone,
-    time: payload.time,
-    comment: payload.comment,
-    payment: payload.payment,
-    delivery: payload.delivery,
-    subtotal: payload.subtotal,
-    extrasTotal: payload.extrasTotal,
-    promoCode: payload.promoCode,
-    promoTitle: payload.promoTitle,
-    promoDiscount: payload.promoDiscount,
-    totalAmount: effectiveTotal,
-    tipPercent: flowType === "sale" ? 0.1 : 1,
-    amountXtr,
-    rental_id: rentalId,
-    startParam,
-    telegramWebappLink,
-    franchizeRentalLink,
-    extras: payload.extras,
-    cartLines: payload.cartLines,
-    flowType,
-  };
-
-  let createdPendingInvoiceForThisAttempt = false;
-
-  try {
-    if (!options.skipNotification) {
-      await buildFranchizeOrderDocAndNotify({ ...payload, totalAmount: effectiveTotal, subtotal: payload.subtotal, extrasTotal: payload.extrasTotal });
-    }
-
-    const invoiceRecord = await createInvoice("franchize_order", invoiceId, payload.telegramUserId, amountXtr, 0, metadata);
-    if (!invoiceRecord.success) {
-      return { success: false, error: invoiceRecord.error ?? "Не удалось создать запись счёта." };
-    }
-    createdPendingInvoiceForThisAttempt = invoiceRecord.data?.status === "pending";
-
-    const invoiceSent = await sendTelegramInvoice(
-      payload.telegramUserId,
-      invoiceTitle,
-      description,
-      invoiceId,
-      amountXtr,
-      0,
-    );
-
-    if (!invoiceSent.success) {
-      if (createdPendingInvoiceForThisAttempt) {
-        await cleanupPendingFranchizeInvoiceAfterSendFailure(invoiceId, rentalId);
-      }
-      return { success: false, error: invoiceSent.error ?? "Не удалось отправить XTR-счёт в Telegram." };
-    }
-
-    return { success: true, invoiceId, amountXtr };
-  } catch (error) {
-    if (createdPendingInvoiceForThisAttempt) {
-      await cleanupPendingFranchizeInvoiceAfterSendFailure(invoiceId, rentalId);
-    }
-    logger.error("[franchize] createFranchizeOrderInvoice failed", error);
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Unexpected invoice error",
-    };
-  }
-}
-
-export async function createFranchizeOrderInvoice(input: unknown): Promise<{ success: boolean; invoiceId?: string; amountXtr?: number; error?: string }> {
-  const parsed = franchizeOrderInvoiceSchema.safeParse(input);
-
-  if (!parsed.success) {
-    return {
-      success: false,
-      error: parsed.error.issues[0]?.message ?? "Некорректный payload для XTR-счёта.",
-    };
-  }
-
-  const totalResult = await resolveFranchizeCheckoutTotal(parsed.data);
-  if (!totalResult.success) return totalResult;
-  return createFranchizeOrderInvoiceInternal({ ...parsed.data, ...totalResult, totalAmount: totalResult.totalAmount });
-}
-
-export async function createFranchizeOrderCheckout(
-  input: unknown,
-): Promise<{ success: boolean; invoiceId?: string; amountXtr?: number; error?: string }> {
-  const parsed = franchizeOrderInvoiceSchema.safeParse(input);
-  if (!parsed.success) {
-    return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный payload заказа." };
-  }
-
-  const payload = parsed.data;
-  const now = Date.now();
-  const throttleKey = `${payload.slug}:${payload.orderId}:${payload.telegramUserId}`;
-  const lastRequestAt = franchizeCheckoutRateLimits.get(throttleKey) ?? 0;
-  if (now - lastRequestAt < FRANCHIZE_CHECKOUT_COOLDOWN_MS) {
-    return { success: false, error: "Слишком частые запросы. Повторите через несколько секунд." };
-  }
-  const totalResult = await resolveFranchizeCheckoutTotal(payload);
-  if (!totalResult.success) return totalResult;
-  franchizeCheckoutRateLimits.set(throttleKey, now);
-  const effectiveTotal = totalResult.totalAmount;
-  const validatedPayload = { ...payload, ...totalResult, totalAmount: effectiveTotal };
-
-  try {
-    await buildFranchizeOrderDocAndNotify({
-      ...validatedPayload,
-      subtotal: payload.subtotal,
-      extrasTotal: payload.extrasTotal,
-      totalAmount: effectiveTotal,
-    });
-
-    if (payload.payment === "telegram_xtr") {
-      return createFranchizeOrderInvoiceInternal(validatedPayload, { skipNotification: true });
-    }
-
-    return { success: true };
-  } catch (error) {
-    logger.error("[franchize] createFranchizeOrderCheckout failed", error);
-    return {
-      success: false,
-      error: error instanceof Error ? error.message : "Unexpected checkout error",
-    };
-  }
-}
-
-const checkFranchizeAvailabilitySchema = z.object({
-  carIds: z.array(z.string().trim().min(1)).min(1),
-  rentalStartDate: z.string().trim().min(1),
-  rentalEndDate: z.string().trim().min(1),
-});
-
-export async function checkFranchizeCarsAvailability(input: unknown): Promise<{ success: boolean; unavailableCarIds?: string[]; error?: string }> {
-  const parsed = checkFranchizeAvailabilitySchema.safeParse(input);
-  if (!parsed.success) {
-    return { success: false, error: parsed.error.issues[0]?.message ?? "Некорректный запрос проверки доступности." };
-  }
-
-  const { carIds, rentalStartDate, rentalEndDate } = parsed.data;
-  const startAt = new Date(`${rentalStartDate}T00:00:00.000Z`);
-  const endAt = new Date(`${rentalEndDate}T23:59:59.999Z`);
-  if (Number.isNaN(startAt.getTime()) || Number.isNaN(endAt.getTime()) || endAt < startAt) {
-    return { success: false, error: "Проверьте диапазон дат аренды." };
-  }
-  const startIso = startAt.toISOString();
-  const endIso = endAt.toISOString();
-
-  const blockingStatuses = ["pending", "pending_confirmation", "confirmed", "active"] as const;
-
-  const { data, error } = await supabaseAdmin
-    .from("rentals")
-    .select("vehicle_id")
-    .in("vehicle_id", carIds)
-    .in("status", blockingStatuses)
-    .filter("requested_start_date", "lt", endIso)
-    .filter("requested_end_date", "gt", startIso);
-
-  if (error) {
-    return { success: false, error: `Не удалось проверить пересечения аренды: ${error.message}` };
-  }
-
-  const unavailableCarIds = Array.from(new Set((data ?? []).map((row) => row.vehicle_id).filter((id): id is string => typeof id === "string")));
-  return { success: true, unavailableCarIds };
-}
-
-export async function getFranchizeRentalCard(slug: string, rentalId: string): Promise<{
-  found: boolean;
-  rentalId: string;
-  slug: string;
-  status: string;
-  paymentStatus: string;
-  totalCost: number;
-  vehicleTitle: string;
-  renterId: string;
-  ownerId: string;
-  metadata: Record<string, unknown> | null;
-  contractVerificationStatus: "verified" | "not_verified" | "expired";
-  contractVerifierScope: string;
-  contractDocumentKey: string;
-  docVerifierRecordId: string;
-  contractSourceScope: string;
-  contractOriginalSha256: string;
-  telegramDeepLink: string;
-}> {
-  const safeSlug = slug.trim();
-  const safeRentalId = rentalId.trim();
-  if (!safeSlug || !safeRentalId) {
-    return {
-      found: false,
-      rentalId: safeRentalId || "-",
-      slug: safeSlug || "vip-bike",
-      status: "pending_confirmation",
-      paymentStatus: "interest_paid",
-      totalCost: 0,
-      vehicleTitle: "—",
-      renterId: "",
-      ownerId: "",
-      metadata: null,
-      contractVerificationStatus: "not_verified",
-      contractVerifierScope: "",
-      contractDocumentKey: "",
-      docVerifierRecordId: "",
-      contractSourceScope: "",
-      contractOriginalSha256: "",
-      telegramDeepLink: `https://t.me/oneBikePlsBot/app?startapp=rental-${safeRentalId}`,
-    };
-  }
-
-  const { data, error } = await supabaseAdmin
-    .from("rentals")
-    .select("rental_id, status, payment_status, total_cost, user_id, owner_id, metadata, vehicle:cars(make, model)")
-    .eq("rental_id", safeRentalId)
-    .maybeSingle();
-
-  if (error || !data) {
-    return {
-      found: false,
-      rentalId: safeRentalId,
-      slug: safeSlug,
-      status: "pending_confirmation",
-      paymentStatus: "interest_paid",
-      totalCost: 0,
-      vehicleTitle: "—",
-      renterId: "",
-      ownerId: "",
-      metadata: null,
-      contractVerificationStatus: "not_verified",
-      contractVerifierScope: "",
-      contractDocumentKey: "",
-      docVerifierRecordId: "",
-      contractSourceScope: "",
-      contractOriginalSha256: "",
-      telegramDeepLink: `https://t.me/oneBikePlsBot/app?startapp=rental-${safeRentalId}`,
-    };
-  }
-
-  const vehicle = data.vehicle as { make?: string; model?: string } | null;
-
-  const metadata = (data.metadata as Record<string, unknown> | null) ?? null;
-  const verifier = metadata && typeof metadata.contract_verifier === "object" ? (metadata.contract_verifier as Record<string, unknown>) : null;
-  const rawStatus = typeof verifier?.status === "string" ? verifier.status.trim().toLowerCase() : "";
-  const expiresAtValue = typeof verifier?.expiresAt === "string" ? verifier.expiresAt : "";
-  const hasExpired = Boolean(expiresAtValue) && new Date(expiresAtValue).getTime() < Date.now();
-  const contractVerificationStatus = hasExpired
-    ? "expired"
-    : rawStatus === "verified" || rawStatus === "ok" || rawStatus === "valid"
-      ? "verified"
-      : rawStatus === "expired"
-        ? "expired"
-      : "not_verified";
-
-  return {
-    found: true,
-    rentalId: data.rental_id,
-    slug: safeSlug,
-    status: data.status ?? "pending_confirmation",
-    paymentStatus: data.payment_status ?? "interest_paid",
-    totalCost: Number(data.total_cost ?? 0),
-    vehicleTitle: `${vehicle?.make ?? "Vehicle"} ${vehicle?.model ?? ""}`.trim(),
-    renterId: data.user_id ?? "",
-    ownerId: data.owner_id ?? "",
-    metadata,
-    contractVerificationStatus,
-    contractVerifierScope: typeof verifier?.scope === "string" ? verifier.scope : "",
-    contractDocumentKey: typeof verifier?.documentKey === "string" ? verifier.documentKey : "",
-    docVerifierRecordId: typeof verifier?.docVerifierRecordId === "string" ? verifier.docVerifierRecordId : "",
-    contractSourceScope: typeof verifier?.sourceScope === "string" ? verifier.sourceScope : "",
-    contractOriginalSha256: typeof verifier?.originalSha256 === "string" ? verifier.originalSha256 : "",
-    telegramDeepLink: `https://t.me/oneBikePlsBot/app?startapp=rental-${data.rental_id}`,
-  };
-}
+// Compatibility surface for existing franchize routes/components.
+// New code should import from the focused modules in `app/franchize/server-actions/*`.
+export type {
+  CatalogItemVM,
+  FranchizeBySlugResult,
+  FranchizeConfigInput,
+  FranchizeConfigState,
+  FranchizeCrewVM,
+  FranchizeHeaderVM,
+  FranchizeTheme,
+  RentalReviewSummaryVM,
+  RentalReviewVM,
+} from "@/app/franchize/actions-runtime";
+
+export { getFranchizeBySlug, markCrewBikesAvailable } from "@/app/franchize/server-actions/catalog";
+export { loadFranchizeConfigBySlug, saveFranchizeConfig } from "@/app/franchize/server-actions/config";
+export { validateFranchizePromoCode } from "@/app/franchize/server-actions/promotions";
+export {
+  createFranchizeOrderCheckout,
+  createFranchizeOrderInvoice,
+  getFranchizeOrderNotificationFailures,
+  retryFranchizeOrderNotification,
+  submitFranchizeOrderNotification,
+} from "@/app/franchize/server-actions/orders";
+export {
+  getFranchizeRentalReviewsForModeration,
+  getRentalReviewContext,
+  moderateRentalReview,
+  submitRentalReview,
+} from "@/app/franchize/server-actions/reviews";
+export {
+  checkFranchizeCarsAvailability,
+  getFranchizeRentalCard,
+  reconcileRentalContractVerifierAttachment,
+} from "@/app/franchize/server-actions/rentals";
